% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{addCollectionMetadata}
\alias{addCollectionMetadata}
\title{Adds metadata to a gene set collection of a GeneSetDb}
\usage{
addCollectionMetadata(x, xcoll, xname, value, validate.value.fn = NULL,
  allow.add = TRUE)
}
\arguments{
\item{x}{\code{GeneSetDb}}

\item{xcoll}{The collection name}

\item{xname}{The name of the metadata variable}

\item{value}{The value of the metadata variable}

\item{validate.value.fn}{If a function is provided, it is run on
\code{value} and msut return \code{TRUE} for addition to be made}

\item{allow.add}{If \code{FALSE}, this xcoll,xname should be in the
\code{GeneSetDb} already, and this will fail because something is
deeply wrong with the world}
}
\value{
The updated \code{GeneSetDb}
}
\description{
This function is exported, but I imagine this being used mostly through
different replace methods that use this as a utility function.
}

