% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{annotateGeneSetMembership}
\alias{annotateGeneSetMembership}
\title{Annotates rows of a data.frame with geneset membership from a GeneSetDb}
\usage{
annotateGeneSetMembership(x, gdb, x.ids = "entrez_id", ...)
}
\arguments{
\item{x}{A data.frame with genes/features in rows}

\item{gdb}{A \code{GeneSetDb} object with geneset membership}

\item{x.ids}{The name of the column in \code{x} that holds the feautre
id's in \code{x} that match the featureId's in \code{gdb}, or a vector
of id's to use for each row in \code{x} that represent these.}
}
\value{
Returns the original \code{x} with additional columns: each is a
  logical vector that indicates membership for genesets defined in
  \code{gdb}.
}
\description{
Annotates rows of a data.frame with geneset membership from a GeneSetDb
}

