% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{append,GeneSetDb,ANY-method}
\alias{append,GeneSetDb,ANY-method}
\title{Appends two GeneSetDb objects togethter.}
\usage{
\S4method{append}{GeneSetDb,ANY}(x, values, after = NA)
}
\arguments{
\item{x}{A \code{\link[=GeneSetDb]{GeneSetDb()}} object}

\item{values}{A \code{\link[=GeneSetDb]{GeneSetDb()}} object}

\item{after}{not used}
}
\description{
Appends two GeneSetDb objects togethter.
}
\examples{
gdb1 <- exampleGeneSetDb()
gdb2 <- GeneSetDb(exampleGeneSetDF())
gdb <- append(gdb1, gdb2)
}
