% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{append,MultiGSEAResult,ANY-method}
\alias{append,MultiGSEAResult,ANY-method}
\title{Appends two MultiGSEAResult objects together.}
\usage{
\S4method{append}{MultiGSEAResult,ANY}(x, values, after = NA)
}
\arguments{
\item{x}{A \code{MultiGSEAResult} object}

\item{values}{Another \code{MultiGSEAResult}}

\item{after}{not used}
}
\value{
A combined \code{MultiGSEAResult} object
}
\description{
This would be useful when you want to add a GSEA result to an already
existing one.
}
\details{
When would you want to do that? Imagine a shiny app that drives multiGSEA.
You might want to present the results of each analysis as they come "online",
so you would run them independently and make them available to the user
immediately after they each finish (ie. in combination with the promises
package).
}
\examples{
gdb1 <- exampleGeneSetDb()
gdb2 <- GeneSetDb(exampleGeneSetDF())
gdb <- append(gdb1, gdb2)
}
