% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{approx.target.from.transformed}
\alias{approx.target.from.transformed}
\title{Get the approximate nominal pvalue for a target qvalue given the
distribution of adjusted pvalues from the nominal ones.}
\usage{
approx.target.from.transformed(target, orig, xformed, thresh = 0.01)
}
\arguments{
\item{thresh}{how close padj has to be in padjs to get its nominal
counterpart}

\item{pval}{the adjust pvalue you want the nominal value for}

\item{pvals}{the distribution of nominal pvalues}

\item{padjs}{the adjusted pvalues from \code{pvals}}
}
\value{
numeric answer, or NA if can't find nominal pvalue within given
  threshold for padjs
}
\description{
Get the approximate nominal pvalue for a target qvalue given the
distribution of adjusted pvalues from the nominal ones.
}
