% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{archiveDir}
\alias{archiveDir}
\title{Retrieve the path to the cache directory.}
\usage{
archiveDir()
}
\value{
String containing the path to the cache directory.
}
\description{
Retrieve the path to the cache directory for the multiGSEA package.
Create the cache directory if need be.
}
