% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{archivePath}
\alias{archivePath}
\title{Retrieve path to a cached file.}
\usage{
archivePath(filename)
}
\arguments{
\item{filename}{Name of the file.}
}
\value{
String containing the path to the file.
}
\description{
The function retrieves the path to a file that is cached in the archive
directory.
}
