% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{as.data.frame.GeneSetDb}
\alias{as.data.frame.GeneSetDb}
\title{Convert a GeneSetDb to other formats.}
\usage{
\method{as.data.frame}{GeneSetDb}(x, row.names = NULL, optional = FALSE,
  value = c("featureId", "x.id", "x.idx"), active.only = is.conformed(x),
  ...)
}
\arguments{
\item{x}{A \code{GeneSetDb} object}

\item{value}{The value to use for the individual feature id's per collection}

\item{active.only}{If the \code{GeneSetDb} is conformed, do you want to only
return the features that match target and are "active"?}

\item{...}{nothing}
}
\value{
a \code{data.frame} of the GeneSetDb
}
\description{
Convert a GeneSetDb to other formats.
}
\section{To a data.frame}{


This will return a data.frame representation of a \code{GeneSetDb} object.
If there is an "active" featureIdMap, it will optionally use that to map
the original featureIds to the new ones specified by the map.
}
\examples{
gdb <- exampleGeneSetDb()
gdf <- as.data.frame(gdb)
head(gdf)
}

