% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-utils.R
\name{as.DGEList}
\alias{as.DGEList}
\title{Utility function to convert ExpressionSet to DGEList}
\usage{
as.DGEList(x, element = "counts", lib.size = NULL, group = rep(1,
  ncol(x)), norm.factors = .pData(x)[["norm.factors"]], ...)
}
\arguments{
\item{x}{ExpressionSet to convert}

\item{element}{The name of the \code{assayDataElement(x)} to use for the
count matrix from \code{x}}

\item{norm.factors}{A numeric vector of norm.factors to use. If this is
NULL, then they will be recalculated}
}
\value{
A DGEList with norm factors
}
\description{
Utility function to convert ExpressionSet to DGEList
}

