% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{as.expression.indexes}
\alias{as.expression.indexes}
\title{Unrolls the GeneSetDb into a list of index vectors per "active" gene set}
\usage{
\method{as.expression}{indexes}(x, value = c("x.idx", "x.id"),
  active.only = TRUE)
}
\arguments{
\item{x}{A GeneSetDb}

\item{value}{\code{x.idx} returns indexes into the conformed expression
object as integers (ie. row index numbers) and \code{x.id} returns them
as their featureId's as used in the target expression object.}
}
\description{
Unrolls the GeneSetDb into a list of index vectors per "active" gene set
}

