% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_matrix}
\alias{as_matrix}
\title{Converts an expression container like object to a matrix for analysis}
\usage{
as_matrix(y)
}
\arguments{
\item{y}{an object to convert into an expression matrix for use in various
internal gene set based methods}
}
\value{
a matrix of values to use downstream of internal gene set based
  methods.
}
\description{
This function is not exported. It will convert various expression-like
containers into a a matrix of values for use in GSEA or single-sample based
scoring methods.
}
\details{
There's nothing too fancy here. Keep in mind, however, that if \code{y}
is something that typically stores counts (a \code{DGEList} or
\code{DESeqDataSet}), then it is transformed into a matrix of values
on the log scale.
}
