% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildMappingTable}
\alias{buildMappingTable}
\title{Builds a table that maps gene set IDs to the data in the expression object x}
\usage{
buildMappingTable(mapping, x, expr.id.fn = rownames)
}
\arguments{
\item{mapping}{A two-column data.frame, with the first column being the
id's of the things used in the genesets, and the second column being the
id's that these things are mapped to in \code{x}}

\item{x}{The "rectangular thing" that holds the epxression data.}
}
\value{
a data.table with the following columns, keyed on "gset.id"
\enumerate{
  \item gset.id The ids used in the genesets
  \item x.id The ids of the same things in \code{x}
  \item x.index The row index of "the thing" in \code{x}
}
}
\description{
Builds a table that maps gene set IDs to the data in the expression object x
}

