% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{check.dt}
\alias{check.dt}
\title{Check a data.table vs a reference data.table}
\usage{
check.dt(x, ref)
}
\arguments{
\item{x}{A \code{data.table} you want to be validated}

\item{ref}{The \code{data.table} to use as the reference/model data.table}
}
\value{
\code{TRUE} if all things check out, otherwise a character vector
  indicating what the problems were.
}
\description{
This function ensures that a \code{data.table} \code{x} has a superset of
columns of a reference table \code{ref}, and that both tables are keyed by
the same columns.
}

