% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{collectionMetadata}
\alias{collectionMetadata}
\title{Query for metadata associated with a collection.}
\usage{
collectionMetadata(x, collection, name)
}
\arguments{
\item{x}{Object to extract the collectionMetadata from}

\item{collection}{The geneset collection to to query}

\item{name}{The name of the metadata variable to get the value for}
}
\description{
Query for metadata associated with a collection.
}

