% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-utils.R
\name{combine.DGEList}
\alias{combine.DGEList}
\title{Combines two (already concordant) DGEList objects}
\usage{
combine.DGEList(x, y)
}
\arguments{
\item{x}{\code{DGEList}}

\item{y}{\code{DGEList}}
}
\value{
combined \code{DGEList}
}
\description{
Combines two (already concordant) DGEList objects
}

