% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_gsea_scatterplot.R
\name{condition_gsea_scatterplot}
\alias{condition_gsea_scatterplot}
\title{Plot a pairwise scatterplot from GSEA data}
\usage{
condition_gsea_scatterplot(
  gsea_result,
  condition1,
  condition2,
  metric = "Directional_Phred"
)
}
\arguments{
\item{gsea_result}{A gsea result from run_multi_gsea or filter_multi_gsea}

\item{condition1}{The name of the first condition}

\item{condition2}{The name of the second condition}

\item{metric}{The metric column to use (NES or Directional_Phred)}
}
\value{
A ggplot object of the pairwise scatterplot between the conditions
}
\description{
Plot a pairwise scatterplot from GSEA data
}
