% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneSetDb-methods.R
\docType{methods}
\name{conform}
\alias{conform}
\alias{conform,GeneSetDb-method}
\title{Map featureId's in a GeneSetDb to a target expression object.}
\usage{
conform(x, ...)

\S4method{conform}{GeneSetDb}(x, target, unique.by = c("none", "mean", "var"),
  min.gs.size = 3L, max.gs.size = Inf, match.tolerance = 0.25, ...)
}
\arguments{
\item{x}{\code{GeneSetDb}}

\item{...}{moar args}

\item{target}{The expression object/matrix to conform to. This could also
just be a character vector of IDs.}

\item{unique.by}{If there are multiple rows that map to the identifiers
used in the genesets, this is a means to pick the single row for that ID}

\item{min.gs.size}{Ensure that the genesets that make their way to the
\code{GeneSetDb@table} are of a minimum size}

\item{max.gs.size}{Ensure that the genesets that make their way to the
\code{GeneSetDb@table} are smaller than this size}

\item{match.tolerance}{Numeric value between [0,1]. If the fraction of
\code{featureId}s used in \code{x} that match \code{rownames(y)} is
below this number, a warning will be fired.}

\item{x}{The GeneSetDb}
}
\value{
A \code{GeneSetDb} that has been matched/conformed to an
  expression object target (\code{y})
}
\description{
Map featureId's in a GeneSetDb to a target expression object.

This uses the already-set featureIdMap for the GeneSetDb
}

