% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-corplot.R
\name{corplot}
\alias{corplot}
\alias{panel.cor}
\alias{panel.hist}
\alias{panel.spoints}
\title{Plots the correlation of genes in signature.}
\usage{
corplot(E, title, cluster = FALSE, col.point = "#00000066",
  diag.distro = TRUE)

panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)

panel.hist(x, ...)

panel.spoints(x, y, col = par("col"), bg = NA, pch = par("pch"),
  cex = 1, col.smooth = "red", span = 2/3, iter = 3,
  col.point = "#00000022", ...)
}
\arguments{
\item{E}{sample-by-gene matrix from \code{\link{fetchExpression}} to plot.
Note that this is transpose of "the usual" gene x sample matrix.}

\item{title}{The title of the plot}

\item{cluster}{\code{logical} indicating whether or not to shuffle genes
around into some clustering.}

\item{col.point}{the color of the points in the scatterplots}

\item{diag.distro}{show the distribution of values on the diagnols?}
}
\value{
nothing, just creates the plot
}
\description{
Plots the correlation of genes in signature.

Plots the correlation stats with colored background for the lower triangle
of pairs plot.

Plots the expression distribution as histogrm (mai diagonal of pair plot)

Scatter plot for upper triangle of pairs plot
}

