% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-utils.R
\name{CPM}
\alias{CPM}
\title{Provides a "regularized" version of edgeR::cpm}
\usage{
CPM(x, normalized.lib.sizes = TRUE, log = TRUE, prior.count = 5,
  prior.smooth = 0.25, regularized = prior.count != prior.smooth,
  norm.factors = NULL, lib.size = NULL, norm.factors.col = "norm.factors",
  lib.size.col = "lib.size", x.element = "counts", ...)
}
\arguments{
\item{x}{A \code{DGEList} or something that can be converted to a count
matrix via \code{as.matrix(x)}}

\item{normalized.lib.sizes}{\code{logical}: use normalized lib sizes?}

\item{log}{\code{log} return expression in log2? (defaults to \code{TRUE})}

\item{prior.count}{The lambda/mean of the Poisson distribution used to add
pseudo counts to if \code{regularized == TRUE}. Otherwise a constant
count added to each observation. Note that a A minimum of
\code{prior.smooth} is always added to the poisson noise to avoid taking
logs of 0}

\item{prior.smooth}{A small number that is added to the Poisson random var
drawn by \code{prior.count} in order to avoid taking logs of 0. This will
always be at least 0.25}

\item{regularized}{If TRUE, prior.count parameterize a Poission distribution
to draw from, otherwise prior.count is constant. When \code{FALSE}, this
function should behave almost identically to \code{\link[edgeR]{cpm}}}

\item{norm.factors}{precalculated norm.factors (numeric vector as long as
\code{ncol(x)})}

\item{lib.size}{precalculate lib.size vector (numeric vector as long as
\code{ncol(x)})}

\item{norm.factors.col}{A column to fetch from \code{.pData(x)} that has
precalculated norm.factors}

\item{lib.size.col}{A column to fetch from \code{.pData(x)} that has
precalculated lib.sizes}

\item{x.element}{Additional param to help get counts out of \code{x}: sent
to \code{.exprs,element} param}

\item{...}{not sure why I have this here}
}
\value{
A \code{matrix,numeric} of the normalized/smoothed expression values
  from \code{x}.
}
\description{
A modified version of edgeR::cpm which uses interprets \code{prior.count}
to parameterize a Poisson distribution to sample read counts from for
each observation
}
\details{
Note that when called with \code{regularized == FALSE}, this
function should behave almost identically to \code{\link[edgeR]{cpm}}.
}

