% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_category_overlap_matrix.R
\name{create_category_overlap_matrix}
\alias{create_category_overlap_matrix}
\title{Calculate the percentage overlap between GO categories}
\usage{
create_category_overlap_matrix(gsea_result, max_categories = 250)
}
\arguments{
\item{gsea_result}{A filtered GSEA result}

\item{max_categories}{Limit to the number of categories to cluster}
}
\value{
A tibble of category pairs with their percentage overlap
}
\description{
Calculate the percentage overlap between GO categories
}
