% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gsea_groups.R
\name{create_gsea_groups}
\alias{create_gsea_groups}
\title{Add grouping information to filtered GSEA results}
\usage{
create_gsea_groups(gsea_matrix, minoverlap = 30)
}
\arguments{
\item{gsea_matrix}{Overlap matrix from create_category_overlap_matrix}

\item{minoverlap}{Minimum percentage overlap to use}
}
\value{
A tibble containing the Cateory IDs and the group to which they were assigned
}
\description{
Add grouping information to filtered GSEA results
}
