% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette-helpers.R
\name{createAirwayDGEList}
\alias{createAirwayDGEList}
\title{Returns a DGEList from the airway dataset for downstream analysis.}
\usage{
createAirwayDGEList()
}
\description{
The airway package provides a dataset that can be used for differential
expression analysis. The object is a SummarizedExperiment with Ensembl
gene IDs as its row identifiers. This function will convert this object
to a DGEList with entrez id's as gene identifiers. Genes with dupliate
entrez ids (or NA ids) will be removed.
}
