% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dir.exists}
\alias{dir.exists}
\title{Utility function to determine if paths listed are real directories}
\usage{
dir.exists(...)
}
\arguments{
\item{...}{character vectors containing file paths. Tilde-expansion is
done: see 'path.expand'.}
}
\value{
A logical vector indicating whether or not the queried directories
exist.
}
\description{
Utility function to determine if paths listed are real directories
}

