% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{dir.writable}
\alias{dir.writable}
\title{Checks that the directory provided is writable by the current user.}
\usage{
dir.writable(path)
}
\arguments{
\item{path}{The path to a directory to check.}
}
\value{
\code{logical}, \code{TRUE} if \code{path} is writable by the
current user, otherise \code{FALSE}
}
\description{
This works by testing to put a temporary file into an already existing
directory
}

