% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateInputs.R
\name{disp.estimated}
\alias{disp.estimated}
\title{Checks a DGEList to see if estimateDisp() was run on it}
\usage{
disp.estimated(x)
}
\arguments{
\item{x}{Input DGEList}
}
\value{
TRUE if yes, FALSE if no or if x is not a DGEList.
}
\description{
Checks a DGEList to see if estimateDisp() was run on it
}

