% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.camera.R
\name{do.camera}
\alias{do.camera}
\title{Worker function to run camera from within a multiGSEA pipeline}
\usage{
do.camera(gsd, x, design, contrast = ncol(design), gs.idxs = as.list(gsd,
  value = "x.idx"), ...)
}
\description{
camera as originally implemented tends to be very conservative, this is
due to the intra-geneset correlation that it originally estimates from
the data. It has been argued that this is hard to estimate and becomes
very conservative for experiments with small N.

As a result, an \code{inter.gene.cor} parameter was introduced to camera
in limma v3.24.14, and as of Bioc3.3, its value is set to 0.01 by default
instead of letting camera approximate it from the data.

For the original discussion around this topic, please refer to the following
thread in the bioconductor support forum:

\url{https://support.bioconductor.org/p/70005/#70195}

\strong{This function is not meant to be called directly, it should only be
called internally within \code{multiGSEA}}
}

