% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.enrich.R
\name{do.enrich}
\alias{do.enrich}
\title{This is a generic wrapper around limma::kegga to perform "biased enrichment"}
\usage{
do.enrich(gsd, x, design, contrast = ncol(design),
  feature.bias = "AveExpr", prior.prob = NULL,
  restrict.universe = FALSE, split.updown = TRUE, use.treat = FALSE,
  feature.min.logFC = if (use.treat) log2(1.25) else 1,
  feature.max.padj = 0.1, logFC = NULL, ...)
}
\arguments{
\item{feature.bias}{we will try to extract the average expression of the
gene as teh default bias, but you can send in gene length, or
what-have-you}
}
\description{
This, in principle, works similarly to goseq but uses \code{\link[limma:kega]{limma::kega()}} as its
engine. If you don't want to calculate any type of biased enrichment, then
explicitly set feature.bias and prior.prob to \code{NULL}.
}
\details{
Running goseq would sometimes throw errors in the \code{makespline} call from
\code{\link[goseq:nullp]{goseq::nullp()}}, so I jumped over to this given this insight:
https://support.bioconductor.org/p/65789/#65914
}
\references{
Young, M. D., Wakefield, M. J., Smyth, G. K., Oshlack, A. (2010).
Gene ontology analysis for RNA-seq: accounting for selection bias.
\emph{Genome Biology} 11, R14. http://genomebiology.com/2010/11/2/R14
}
