% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.fgsea.R
\name{do.fgsea}
\alias{do.fgsea}
\title{Runs GSEA on a pre-ranked list of differential expression statistcis with fgsea}
\usage{
do.fgsea(
  gsd,
  x,
  design,
  contrast = ncol(design),
  sampleSize = 101,
  eps = 1e-10,
  scoreType = c("std", "pos", "neg"),
  nproc = 0,
  gseaParam = 1,
  nPermSimple = 1000,
  absEps = NULL,
  use.fgsea.simple = FALSE,
  score.by = c("t", "logFC", "pval"),
  logFC = NULL,
  gs.idxs = as.list(gsd, active.only = TRUE, value = "x.idx"),
  ...
)
}
\arguments{
\item{gsd}{The \code{\link[=GeneSetDb]{GeneSetDb()}} for analysis}

\item{x}{The expression object. This can be 1 column matrix if you are not
running any analysis, and this function essentially is just a
"pass through"}

\item{design}{The design matrix for the experiment}

\item{contrast}{The contrast you want to test and provide stats for. By
default this tests the last column of the \code{design} matrix. If you
want to test a custom contrast, this can be a contrast vector, which
means that it should be as long as \code{ncol(design)} and it most-often sum to
one. In the future, the user will be able to specify a range of
coefficients over \code{design} to perform an ANOVA analysis, cf.
Issue #11 (https://github.com/lianos/multiGSEA/issues/11).}

\item{...}{arguments to pass down into \code{\link[=calculateIndividualLogFC]{calculateIndividualLogFC()}}}
}
\value{
A data.table of fgsea results.
}
\description{
fgsea expects the pathways to be describes as a list of character vectors
where the vectors are gene ids, and the names of the pre-ranked vector
correspond to those IDs
}
\details{
Deafults fgsea functionality was changed to \code{\link[fgsea:fgseaMultilevel]{fgsea::fgseaMultilevel()}} in
1.13.2 and the default parameters here reflect that. If you want to use the
original "fgseaSimple" you have to pass in \code{use.fgsea.simple = TRUE} in your
\code{multiGSEA()} call.

\code{minSize} and \code{maxSize} are already set by the \code{conform} logic that was
specified in the call to \code{\link[=multiGSEA]{multiGSEA()}} via the \code{min.gs.size} and
\code{max.gs.size} parameters.

\strong{This function is not meant to be called directly.} It should only be
called internally within \code{\link[=multiGSEA]{multiGSEA()}}.
}
