% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.fry.R
\name{do.fry}
\alias{do.fry}
\title{Worker function to run fry from within a multiGSEA pipeline}
\usage{
\method{do}{fry}(gsd, x, design, contrast = ncol(design),
  gs.idxs = as.list(gsd, active.only = TRUE, value = "x.idx"), ...)
}
\description{
\strong{This function is not meant to be called directly, it should only be
called internally within \code{multiGSEA}}
}
