% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.geneSetTest.R
\name{do.geneSetTest}
\alias{do.geneSetTest}
\title{Worker function to run geneSetTest from within a multiGSEA pipeline}
\usage{
\method{do}{geneSetTest}(gsd, x, design, contrast = ncol(design),
  score.by = c("t", "logFC", "pval"), logFC = NULL,
  gs.idxs = as.list(gsd, active.only = TRUE, value = "x.idx"), ...)
}
\description{
\strong{This function is not meant to be called directly, it should only be
called internally within \code{multiGSEA}}
}
