% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.goseq.R
\name{do.goseq}
\alias{do.goseq}
\title{Performs goseq analysis significance of gene set membership.}
\usage{
do.goseq(gsd, x, design, contrast = ncol(design), feature.bias,
  goseq.method = "Wallenius", repcnt = 2000,
  use_genes_without_cat = TRUE, split.updown = TRUE,
  direction = c("over", "under"), plot.fit = FALSE,
  use.treat = FALSE, feature.min.logFC = if (use.treat) log2(1.25) else
  1, feature.max.padj = 0.1, logFC = NULL, ...)
}
\arguments{
\item{gsd}{The \code{\link{GeneSetDb}} for analysis}

\item{x}{The expression object}

\item{design}{Experimental design}

\item{contrast}{The contrast to test}

\item{feature.bias}{a named vector as long as \code{nrow(x)} that has the
"bias" information for the features/genes tested (ie. vector of gene
lengths). \code{names(feature.bias)} should equal \code{rownames(x)}.
The caller MUST provide this. The goseq package provides a
\code{\link[goseq]{getlength}} function which facilitates getting default
values for these if you do not have the correct values used in your
analysis. If there is no way for you to get this information, then use
enrichtest with no \code{feature.bias} vector.}

\item{repcnt}{Number of random samples to be calculated when random sampling
is used. Ignored unless \code{method="Sampling"}.}

\item{use_genes_without_cat}{A boolean to indicate whether genes without a
categorie should still be used. For example, a large number of gene may
have no GO term annotated. If this option is set to FALSE, those genes
will be ignored in the calculation of p-values (default behaviour). If
this option is set to TRUE, then these genes will count towards the total
number of genes outside the category being tested.}

\item{direction}{Same as direction in \code{GOstats}}

\item{plot.fit}{To plot (or not) the bias in selected genes vs.
\code{feature.bias}.}

\item{logFC}{The logFC data.table from \code{calculateIndividualLogFC}}

\item{...}{arguments to pass down into \code{calculateIndividualLogFC}}

\item{method}{The method to use to calculate the unbiased category
enrichment scores}
}
\value{
A data.table of goseq results. The "pval" column here refers to
pval.over, for simplicity in other places. If \code{split.updown=TRUE},
a list of data.table's are returned named 'goseq', 'goseq.up', and
'goseq.down' which are the results of running goseq three independent
times.
}
\description{
Genes are selected for testing against each geneset by virture of them
passing a maximum FDR and minimum log fold change as perscribed by the
\code{min.logFC} and \code{max.padj} parameters, respectfully.
}
\details{
Note that we are intentionally adding a hyperG.selected column by reference
so that this information is kicked back to the caller multiGSEA function
and included in downstream reporting.

\strong{This function is not meant to be called directly.} It should only be
called internally within \code{\link[=multiGSEA]{multiGSEA()}}.
}
