% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.GSEA-P.R
\name{do.gseap}
\alias{do.gseap}
\title{Runs a method analagous to GSEA-P}
\usage{
do.gseap(gsd, x, design, contrast, logFC = NULL, score.by = c("t", "logFC"),
  robust.fit = FALSE, robust.eBayes = FALSE, ...)
}
\description{
See the following link for reference
http://guangchuangyu.github.io/2015/11/comparison-of-clusterprofiler-and-gsea-p/
}

