% Generated by roxygen2 (4.0.2): do not edit by hand
\name{do.hyperGeometricTest}
\alias{do.hyperGeometricTest}
\title{Performs a hypergeometric test for significance of gene set membership.}
\usage{
do.hyperGeometricTest(x, gs.table, design, contrast, direction = "over",
  logFC.stats = NULL, robust.fit = FALSE, robust.eBayes = FALSE,
  min.logFC = 1, max.padj = 0.1, ...)
}
\arguments{
\item{x}{The expression object}

\item{gs.table}{The \code{\link{GeneSetTable}} for analysis}

\item{design}{Experimental design}

\item{contrast}{The contrast to test}

\item{direction}{Same as direction in \code{GOstats}}

\item{logFC.stats}{The logFC stats table for the given contrast}

\item{logFC.stats}{The topTable from limma}
}
\description{
Genes are selected for testing against each geneset by virture of them
passing a maximum FDR and minimum log fold change as perscribed by the
\code{min.logFC} and \code{max.padj} parameters, respectfully.
}

