% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/do.logFC.R
\name{do.logFC}
\alias{do.logFC}
\title{A "pass through" function. The caller asks for "logFC" if she only wants
to calculate the geneSet statistics for each of the geneSets, ie. their
JG, mean.logFC, mean.t for each geneset.}
\usage{
do.logFC(gsd, x, design, contrast = ncol(design), outdir = NULL,
  use.cache = TRUE, logFC = NULL, robust.fit = FALSE,
  robust.eBayes = FALSE, ...)
}
\description{
The user can get all the relenvant stats for this by calling either
\code{logFC(MultiGSEAResult} to get the individual logFCs, or
\code{geneSets(MultiGSEAResult} to get the geneSet level statistics like
the JG score.
}

