% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.roast.R
\name{do.roast}
\alias{do.roast}
\title{Worker function to run roast from within a multiGSEA pipeline}
\usage{
do.roast(gsd, x, design, contrast = ncol(design), gs.idxs = as.list(gsd,
  active.only = TRUE, value = "x.idx"), ...)
}
\description{
\strong{This function is not meant to be called directly, it should only be
called internally within \code{multiGSEA}}
}
