% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.romer.R
\name{do.romer}
\alias{do.romer}
\title{Worker function to run romer from within a multiGSEA pipeline}
\usage{
\method{do}{romer}(gsd, x, design, contrast = ncol(design),
  gs.idxs = as.list(gsd, active.only = TRUE, value = "x.idx"), ...)
}
\description{
Note that romer works on a DGEList or a "normal" EList, ie. it has not
been updated to work with an \code{EList} with a \code{$weights} matrix,
and therefore doesn't work with a voom'd dataset.

\strong{This function is not meant to be called directly, it should only be
called internally within \code{multiGSEA}}
}
