% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score-genesets.R
\name{do.scoreGeneSets.zscore}
\alias{do.scoreGeneSets.zscore}
\title{Default to sqrt in denominator of zscores to stabilize the variance of
the mean:}
\usage{
do.scoreGeneSets.zscore(gdb, y, zsummary = c("sqrt", "mean"), trim = 0.1,
  melted = FALSE, ...)
}
\description{
Lee, E., et al. Inferring pathway activity toward precise disease
classification. PLoS Comput. Biol. 4, e1000217 (2008).
}

