% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{do.scoreSingleSamples.gsdecon}
\alias{do.scoreSingleSamples.gsdecon}
\title{Jason's method}
\usage{
do.scoreSingleSamples.gsdecon(gdb, y, as.matrix = FALSE, design = NULL,
  doPerm = FALSE, nPerm = 249, pvalueCutoff = 0.01, nComp = 1,
  seed = NULL, ...)
}
\description{
Jason's method
}

