% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{do.scoreSingleSamples.svd}
\alias{do.scoreSingleSamples.svd}
\title{A no dependency call to GSDecon's eigengene scoring}
\usage{
do.scoreSingleSamples.svd(gdb, y, as.matrix = FALSE, center = TRUE,
  scale = TRUE, uncenter = center, unscale = scale, ...)
}
\description{
A no dependency call to GSDecon's eigengene scoring
}
