% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.svdGeneSetTest.R
\name{do.svdGeneSetTest}
\alias{do.svdGeneSetTest}
\title{Transforms gene x sample to geneset x sample and run differential expression}
\usage{
do.svdGeneSetTest(gsd, x, design, contrast = ncol(design),
  gs.idxs = as.expression.indexes(gsd, value = "x.idx"),
  trend.eBayes = FALSE, ...)
}
\description{
Transforms gene x sample to geneset x sample and run differential expression
}

