% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.enrich.R
\name{enrichment}
\alias{enrichment}
\title{Runs biased enrichment test over a data.frame of features}
\usage{
enrichment(gsd, dat, selected = "significant", groups = "direction",
  feature.bias = NULL, universe = NULL, restrict.universe = FALSE,
  plot.bias = FALSE, ...)
}
\arguments{
\item{gsd}{The GeneSetDb}

\item{dat}{A data.frame with feature-level statistics. Minimally, this should
have a \code{"featureId"} (character) column, but read on ...}

\item{selected}{Either the name of a logical column in \code{dat} used to subset
out the features to run the enrichement over, or a character vector of
\code{"featureId"}s that are selected from \code{dat[["featureId"]]}.}

\item{groups}{Encodes groups of features that we can use to test selected
features individual, as well as "all" together. This can be specified by:
(1) specifying a name of a column in \code{dat} to split the enriched features
into subgroups. (2) A named list of features to intersect with \code{selected}.
By default this is \code{NULL}, so we only run enrichment over
all elements in \code{selected}. See examples for details.}

\item{feature.bias}{If \code{NULL} (default), no bias is used in enrichment
analysis. Otherwise, can be the name of a column in \code{dat} to extract
a numeric bias vector (gene length, GC content, average expression, etc.)
or a named (using featureIds) numeric vector of the same.}

\item{universe}{Defaults to all elements in \code{dat[["featureId"]]}.}

\item{restrict.universe}{See same parameter in \code{\link[limma:kegga]{limma::kegga()}}}

\item{plot.bias}{See \code{plot} parameter in \code{\link[limma:kega]{limma::kega()}}}
}
\value{
a data.frame of pathway enrichment
}
\description{
This function wraps \code{\link[limma:kegga]{limma::kegga()}} to perform biased enrichment tests over
gene set collection stored in a GeneSetDb (\code{gsd}) object. Its easiest to
use this function when the biases and selection criteria are stored as
columns of the input data.frame \code{dat}.
}
