% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exampleGeneSets}
\alias{exampleGeneSets}
\title{Get sample gene sets}
\usage{
exampleGeneSets(as = c("lol", "limma"))
}
\arguments{
\item{as}{Character vector to specify the type of returned geneset
list that you want.}
}
\value{
A list of lists of entrezIDs when \code{as == 'lol'}, or
a list of integers into the rows of \code{exampleExpressionSet}
for the genes in the given geneset.
}
\description{
Get sample gene sets
}

