% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{exampleExpressionSet}
\alias{exampleExpressionSet}
\alias{exampleGeneSets}
\alias{exampleGeneSetDb}
\alias{exampleGeneSetDF}
\title{Functions that load data for use in examples and testing.}
\usage{
exampleExpressionSet(dataset = c("tumor-vs-normal", "tumor-subtype"),
  do.voom = TRUE)

exampleGeneSets(x, unlist = !missing(x))

exampleGeneSetDb()

exampleGeneSetDF()
}
\arguments{
\item{dataset}{Character vector indicating what samples wanted, either
\code{"tumor-vs-normal"} for a tumor vs normal dataset from TCGA, or
just the tumor samples from the same annotated with subtype.}

\item{do.voom}{If TRUE, a voomed EList is returned, otherwise an
ExpressionSet of counts.}

\item{x}{If provided, an expression/matrix object so that the genesets are
returned as (integer) index vectors into the rows of x whose rownames
match the ids in the geneset.}
}
\value{
A list of lists of entrezIDs when \code{as == 'lol'}, or
  a list of integers into the rows of \code{exampleExpressionSet}
  for the genes in the given geneset.
}
\description{
We provide examplar expression data (counts or voomed) as well as exemplar
gene sets in different forms.
}
\section{exampleExpressionSet}{

The expression data is a subset of the TCGA BRCA indication. Calling
\code{exampleExpressionSet} will load either \code{limma} or \code{Biobase}
into the calling environment so that the user can manipulate the object
that is returned.
}

\section{exampleGeneSets}{

Returns gene sets as either a list of feature identifiers or integers
that index into a target expression object \code{x}.
}

\section{exampleGeneSetDb}{

Returns gene sets as a \code{GeneSetDb} object
}

\section{exampleGeneSetDF}{

Returns a data.frame of gene set definitions. A data.frame of this form
can be passed into the \code{GeneSetDb} contructor.
}

