% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gsea_genes.R
\name{extract_gsea_genes}
\alias{extract_gsea_genes}
\title{Extract the leading edge genes from GSEA to a tibble}
\usage{
extract_gsea_genes(gsea_result, max_genes = 100)
}
\arguments{
\item{gsea_result}{A full multi_gsea result tibble}
}
\value{
A tibble containing the extracted genes
}
\description{
Extract the leading edge genes from GSEA to a tibble
}
