% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{ncbi.entrez.link}
\alias{ncbi.entrez.link}
\alias{genecards.entrez.link}
\title{Transforms a column in feature table to an external link for that feature.}
\usage{
ncbi.entrez.link(x, link.col = "symbol")

genecards.entrez.link(x, link.col = "symbol")
}
\arguments{
\item{x}{a data.frame from \code{logFC(MultiGSEAResult)}}

\item{link.col}{the column in \code{x} that should be transformed to a link}
}
\value{
a modified \code{x} with an html link in \code{link.col}
}
\description{
When listing features in an interactive table, it's often useful to link
the feature to an external webpage that has more information about that
feature. Functions to genes to their NCBI or GeneCards webpage via their
\code{featureId} are provided via \code{ncbi.entrez.link} and
\code{genecards.entrez.link}. The column used to transform into a link
is specified by \code{link.col}.

If \code{link.col} is not found in the data.frame \code{x} then the provided
functions are NO-OPS, ie. the same data.frame is simply returned.
}
