% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\docType{methods}
\name{featureIdMap<-,GeneSetDb-method}
\alias{featureIdMap<-,GeneSetDb-method}
\title{Replacing the featureIdMap blows away the "conform"-ation status of x}
\usage{
\S4method{featureIdMap}{GeneSetDb}(x) <- value
}
\description{
This method ensures that there is only one featureId <-> x.id mapping value.
Note that this does not mean that this enforces a 1:1 mapping, it's just
that the same mapping is not listed more than once.
}

