% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_multi_gsea.R
\name{filter_multi_gsea}
\alias{filter_multi_gsea}
\title{Filter a multi GSEA result}
\usage{
filter_multi_gsea(gsea_result, nes_cutoff = 1, fdr_cutoff = 0.05)
}
\arguments{
\item{gsea_result}{The output of run_multi_gsea}

\item{nes_cutoff}{The absolute cutoff value for the NES value}

\item{fdr_cutoff}{The cutoff for the corrected pvalue}
}
\description{
This function filters the output of a previous run of run_multi_gsea.
You set a cutoff on both the (absolute) NES value, and the corrected
pvalue.  Only Gene Ontology IDs which pass both of those filters
in any of the experimental Categories are retained in the returned
results
}
