% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.GSEA.plots}
\alias{generate.GSEA.plots}
\title{Generates GSEA plots for the genesets over a given experiment.}
\usage{
generate.GSEA.plots(x, design, contrast, result, outdir, use.cache = TRUE,
  logFC = NULL, padj.threshold = 1)
}
\arguments{
\item{x}{The expression object used in the multiGSEA call}

\item{design}{The desing matrix for this GSEA}

\item{contrast}{The contrast being analyzed in the GSEA}

\item{result}{The result data.table returned (or being generated) in the
\code{multiGSEA} call.}

\item{outdir}{The path to the parent directory that holds the output for
this run.}

\item{use.cache}{If true, the function checks if the image already exists
before replotting it.}

\item{logFC}{pre-computed logFCs for the contrast we are running GSEA
against.}

\item{padj.threshold}{plot generation to only include genesets that
have some value in an FDR column that is less than or equal to this
threshold.}
}
\value{
A data.table that has the group, id, and path columns. path
is the path to the image for the given geneset.
}
\description{
Generates GSEA plots for the genesets over a given experiment.
}

