% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{geneSetCollectionURLfunction}
\alias{geneSetCollectionURLfunction}
\title{exportMethod geneSetCollectionURLfunction}
\usage{
geneSetCollectionURLfunction(x, i, ...)
}
\description{
exportMethod geneSetCollectionURLfunction
}

