% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-geneSetContrastView.R
\name{geneSetContrastViewUI}
\alias{geneSetContrastViewUI}
\alias{geneSetContrastView}
\alias{is.geneSetContrastViewer}
\alias{updateActiveGeneSetInContrastView}
\title{A module to encapsulate browsing differential statistics of a geneset.}
\usage{
geneSetContrastViewUI(id, height = "590px", width = "400px")

geneSetContrastView(input, output, session, mgc, server = TRUE,
  maxOptions = Inf, sep = "_::_", feature.link.fn = ncbi.entrez.link,
  itools = c("wheel_zoom", "box_select", "reset", "save"))

is.geneSetContrastViewer(x)

updateActiveGeneSetInContrastView(session, viewer, geneset, mgc)
}
\arguments{
\item{id}{the shiny id of the module}

\item{height, width}{the height and width of the module}

\item{input}{the shiny-required bits for the module}

\item{output}{the shiny-required bits for the module}

\item{session}{the shiny-required bits for the module}

\item{mgc}{A \code{\link{MultiGSEAResultContainer}} object}

\item{server}{boolean to indicate whether the genesets in the geneSetSelect
widget should be rendered server side or not (Default: \code{TRUE})}

\item{maxOptions}{a paremeter used to customize the
\code{GeneSetSelect::selectizeInput} UI element. I thought one might want
to tweak this, but I just leave it as is.}

\item{sep}{the separater to put between the collection and name bits of a
geneset. These are the values used in the gene set \code{selectizeInput}.}
}
\value{
\code{geneSetContrastViewUI} returns tagList of html stuff to dump
  into the UI.

the \code{geneSetContrastView} module returns a reactive list,
  with a \code{$gs} element that indicates the currently active geneset in
  the `geneSetSelect` module, and a \code{$selected} element, a character
  vector of feature_ids currently brushed in a contrast view.
}
\description{
The module is meant to be displayed in "a box" so that the user can examine
the coherent (or not) behavior of the geneset across a contrast with respect
to the background distribution of all genes in the contrast.
}
\details{
Embedded within this module is the \code{\link{geneSetSelect}} module, which
provides the list of genesets the user can examine, as well as the title of
the current geneset under scrutiny.

Below the geneSet picker, we embed an \code{\link{iplot}} so that the
user can observe the behavior of the geneset across the contrast. The user
can pick the type of plot to show (density or boxplot) as well as which
statistics to use for plotting (logFC or t-statistics).

A \code{updateActiveGeneSetInContrastView} function is provided to enable
interactions external to this module the ability to update the geneset
selected in the \code{\link{geneSetSelect}} module.
}
