% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GeneSetDb-class}
\alias{.GeneSetDb}
\alias{GeneSetDb-class}
\title{A class that holds a data.table with geneset membership.}
\description{
A class that holds a data.table with geneset membership.
}
\section{Slots}{

\describe{
\item{\code{table}}{The "gene set table": a data.table with geneset information.
Columns include collection, name, N, and n. The end user can add more
columns to this data.table as desired. The actual feature IDs are
computed on the fly by doing a db[J(group, id)]}

\item{\code{db}}{A \code{data.table} to hold all of the original geneset id
information that was used to construct this \code{GeneSetDb}.}

\item{\code{featureIdMap}}{Maps the ids used in the geneset lists to the ids
(rows) over the expression data the GSEA is run on}

\item{\code{collectionMetadata}}{A \code{data.table} to keep metadata about each
individual geneset collection, ie. the user might want to keep track of
where the geneset definitions come from. Perhaps a function that parses
the collection,name combination to generate an URL that points the user
to more information about that geneset, etc.}
}}

