% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{geneSetDb}
\alias{geneSetDb}
\title{Get the GeneSetDb from MultiGSEAResult}
\usage{
geneSetDb(x)
}
\arguments{
\item{x}{\code{MultiGSEAResult}}
}
\value{
The \code{GeneSetDb}
}
\description{
Get the GeneSetDb from MultiGSEAResult
}

