% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{geneSetFeatureStatistics}
\alias{geneSetFeatureStatistics}
\title{Summarize the statistics for each feature at the geneset level}
\usage{
geneSetFeatureStatistics(x, feature.min.logFC = 1, feature.max.padj = 0.1,
  trim = 0.1, ...)
}
\arguments{
\item{x}{A \code{multiGSEAResult} object}

\item{feature.min.logFC}{used with \code{feature.max.padj} to identify
the individual features that have a significant change.}

\item{feature.max.padj}{used with \code{feature.min.logFC} to identify
the individual features that have a significant change.}

\item{trim}{passed to \code{geneSetFeatureStatistics} for trimmed mean
  calculations.}
}
\value{
A data.table with statistics on the effect size shift for the gene
  set as well as numbers of members that shift up or down.
}
\description{
Summarize the statistics for each feature at the geneset level
}

