% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{geneSets}
\alias{geneSets}
\title{Fetch the gene sets "active" (or otherwise) in the GeneSetDb or Result}
\usage{
geneSets(x, ...)
}
\arguments{
\item{x}{GeneSetDb}
}
\value{
a data.table with geneset information.
}
\description{
Fetch the gene sets "active" (or otherwise) in the GeneSetDb or Result
}

