% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{geneSetsStats}
\alias{geneSetsStats}
\title{Summarizes useful statistics per gene set from a MultiGSEAResult}
\usage{
geneSetsStats(x, feature.min.logFC = 1, feature.max.padj = 0.1,
  trim = 0.1, .external = TRUE)
}
\arguments{
\item{x}{A \code{multiGSEAResult} object}

\item{feature.min.logFC}{used with \code{feature.max.padj} to identify
the individual features that are to be considered differentially
expressed.}

\item{feature.max.padj}{used with \code{feature.min.logFC} to identify
the individual features that are to be considered differentially
expressed.}

\item{trim}{The amount to trim when calculated trimmed \code{t} and
\code{logFC} statistics for each geneset.}
}
\value{
A data.table with statistics on the effect size shift for the gene
  set as well as numbers of members that shift up or down.
}
\description{
This function calculates the number of genes that move up/down for the
given contrasts, as well as mean and trimmed mean of the logFC and
t-statistics. Note that the statistics calculated and returned here are
purely a function of the statistics generated at the gene-level stage
of the analysis.
}
\examples{

vm <- exampleExpressionSet(do.voom=TRUE)
gdb <- exampleGeneSetDb()
mg <- multiGSEA(gdb, vm, vm$design, 'tumor', methods=NULL)
head(geneSetsStats(mg))
}

