% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{geneSetStats}
\alias{geneSetStats}
\title{Fetch the logFC statistics for a geneset}
\usage{
geneSetStats(x, i, j)
}
\arguments{
\item{x}{\code{MultiGSEAResult}}

\item{i}{The collection name}

\item{j}{The gene set name}
}
\value{
data.table with the stats for the features in the geneset
}
\description{
Fetch the logFC statistics for a geneset
}

