% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{geneSetSummaryByGenes}
\alias{geneSetSummaryByGenes}
\title{Creates a geneset summary table for genesets defined by gene membership}
\usage{
geneSetSummaryByGenes(x, features, with.features = TRUE,
  feature.rename = NULL, ...)
}
\arguments{
\item{x}{A GeneSetDb or a multiGSEA ResultmultiGSEA result}

\item{features}{The featureIds to use to select genesets from \code{x}}

\item{with.features}{Inlcude the features as columns of the outgoing result?
(Default: \code{TRUE})}

\item{feature.rename}{If features are added as columns, then we might want
to rename them from featureIds to something else. By default these columns
are prefixed with "featureId_", unless \code{feature.rename=FALSE}}

\item{name}{If specified, one of \code{\link{resultNames}} from the
\code{MultiGSEAResult} object \code{x} to use as a look up and filter
geneset by.}

\item{max.p}{The FDR threshold to filter geneset results by in combination
with the \code{name} parameter.}
}
\value{
A summary data.frame
}
\description{
Creates a geneset summary table for genesets defined by gene membership
}

