% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GeneSetTable-class}
\alias{GeneSetTable-class}
\title{A class that holds a data.table with geneset membership.}
\description{
A class that holds a data.table with geneset membership.
}
\section{Slots}{

\describe{
\item{\code{table}}{The data.table with geneset information}

\item{\code{db}}{Not used yet, but putting it here for the future. This will hold
ALL geneset feature membership information, irrespective of whether or not
that feature appears in the expression object that is being tested
against.}

\item{\code{feature.lookup}}{Maps the ids used in the geneset lists to the ids
(rows) over the expression data the GSEA is run on}

\item{\code{species}}{Either "human" or "mouse" -- only (really) used when the object
was built from a call to MSigDB}
}}

