% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GeneSetTable}
\alias{GeneSetTable}
\title{Create a GeneSetTable to match an input expression object.}
\usage{
GeneSetTable(gene.sets, x, mapping = NULL, min.gs.size = 5,
  stop.on.duplicate.x.xref = TRUE, unique.by = c("mean", "var", "none"),
  species = "human", version = .wehi.msigdb.current)
}
\arguments{
\item{gene.sets}{This can be several things: (a) a character vector that
lists the MSigDB set id's (in conjunction with the \code{species}
parameter); (b) a list of logical vectors into the rows of \code{x}. This
should be a list of lists, where the top-level lists identified the "group"
its child elements (genesets) belong to (like the c1, c2, etc. grouping)
in the MSigDB lists; ie. the provenance of the gene set (c) An already
rigged up GeneSetTable;}

\item{x}{The expression values measured. Could be a matrix, ExpressionSet,
etc. rownames are required on this "thing" for it to work.}

\item{mapping}{A data.frame with two columns. The first column is the ids
used to identify the genes in \code{gene.sets}, the second column are the
rows of x that match the particular gene.set id. This is minimally meant to
support GSEA over microarray datsets, where the geneset IDs are entrez IDs,
but \code{rownames(x)} are probeset IDs.}
}
\description{
Create a GeneSetTable to match an input expression object.
}
\section{Easily mapping rownames(x) to IDs used in gene.sets}{


It may be that the IDs used in the provided \code{gene.sets} are different
than the ones used in \code{rownames(x)}, for instance the IDs in
\code{gene.sets} might (are likely to) be Entrez IDs, and the rownames of
\code{x} might by affy probe set IDs, what then?

This is where the \code{mapping} parameter becomes useful.
TODO: document the GeneSetTable ID mapping more thoroughly.
}

