% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overlap.R
\name{get_overlap}
\alias{get_overlap}
\title{Find the overlap between the genes in two GO cateogories}
\usage{
get_overlap(gsea_genes, ID1, ID2)
}
\arguments{
\item{gsea_genes}{a gene set from extract_gsea_genes}

\item{ID1}{The first GO category ID}

\item{ID2}{The second GO category ID}
}
\value{
The average percentage overlap between the two categories
}
\description{
Find the overlap between the genes in two GO cateogories
}
