% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{getGOslimGeneSetDb}
\alias{getGOslimGeneSetDb}
\title{Return the GO slim annotations}
\usage{
getGOslimGeneSetDb(species = c("human", "mouse"))
}
\arguments{
\item{species}{"human" or "mouse"}
}
\value{
\code{GeneSetDb} of the GO slim mappings
}
\description{
\href{http://geneontology.org/page/go-slim-and-subset-guide}{GO Slims} are
"cut down" versions of the GO ontology that contain a subset of the terms in
the whole GO.
}
\details{
PANTHER provides their own set of
\href{GO slims}{http://www.pantherdb.org/panther/ontologies.jsp}, although
it's not clear how often these get updated.
}

