% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-housekeeping.R
\name{getHousekeepingGeneSetDb}
\alias{getHousekeepingGeneSetDb}
\title{Retrieves a housekeeping GeneSetDb for a given species}
\usage{
getHousekeepingGeneSetDb(
  species = c("human", "mouse"),
  id.type = c("entrez", "ensembl"),
  ...
)
}
\arguments{
\item{species}{\code{"human"} or \code{"mouse"}}

\item{id.type}{defines the type of identifers used in the \code{feature_id} column
for the GeneSetDb, \code{"entrez"} or \code{"ensembl"}.}
}
\value{
a GeneSetDb of housekeeping genes
}
\description{
The genesets provided here were extracted from the genes used in the
Qiagen RT2 Profiler PCR Arrays.
}
\details{
Sources for the mouse and human genesets can be found here:
\itemize{
\item mouse: https://www.qiagen.com/us/shop/pcr/primer-sets/rt2-profiler-pcr-arrays/?catno=PAMM-000Z#geneglobe
\item human: https://www.qiagen.com/us/shop/pcr/primer-sets/rt2-profiler-pcr-arrays/?catno=PAHS-000Z#geneglobe
}
}
