% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getIDMappingDatabase}
\alias{getIDMappingDatabase}
\title{Get the correct ID mapping database}
\usage{
getIDMappingDatabase(organism)
}
\arguments{
\item{organism}{String that defines the organism.}
}
\value{
AnnotationDbi database for ID mapping.
}
\description{
Check by means of the given organism name if the required `AnnotationDbi`
package is installed. Select the ID mapping table based on the organism name
and return it.
}
