% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_processing.R
\name{getMetaboliteIDformats}
\alias{getMetaboliteIDformats}
\title{Helper function to get all different metabolite ID formats}
\usage{
getMetaboliteIDformats(pathways)
}
\arguments{
\item{pathways}{List of pathway databases and their pathway definition.}
}
\value{
List of metabolite ID formats.
}
\description{
This helper function extracts all used ID formats in all pathways
and returns a nested list for each pathway database.
}
