% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msigdb-wehi.R
\name{getMSigDBset}
\alias{getMSigDBset}
\title{Fetches the requested MSigDB gene sets.}
\usage{
getMSigDBset(id, species = "human", as.list = FALSE,
  version = .wehi.msigdb.current)
}
\arguments{
\item{id}{A character vector (or integer vector) indicating the desired
genesets}

\item{species}{Character vector indicating which species desired,
\code{"human"} or \code{"mouse"}}

\item{as.list}{If \code{TRUE} returns a list of lists of entrez IDs for
each gene set, otherwse will return a \code{GeneSetDb}}

\item{version}{The version of the MSigDB desired. Defaults to latest.}
}
\value{
A list of genesets, named by \code{id}.
}
\description{
Fetches the requested MSigDB gene sets.
}

