% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_processing.R
\name{getOrganisms}
\alias{getOrganisms}
\title{Get list of supported organisms}
\usage{
getOrganisms()
}
\value{
List of supported organisms
}
\description{
Get a list of organisms that are covered in our workflow through a supporting
`AnnotationDBi` package. Without such a package we would not be able to map
transcript and protein identifier between different formats. All the
organisms that are listed here have at lest kegg and or reactome pathway
annotation that can be queried by `graphite`.
}
\examples{
getOrganisms()

}
