% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.goseq.R
\name{goseq.gene2cat}
\alias{goseq.gene2cat}
\title{Takes a conformed GeneSetDb and creates a data.frame for goseq to test over}
\usage{
goseq.gene2cat(gsd)
}
\arguments{
\item{gsd}{GeneSetDb to prep}
}
\value{
\code{data.frame} for goseq's \code{gene2cat}
}
\description{
Takes a conformed GeneSetDb and creates a data.frame for goseq to test over
}

