% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{.gsd.row.index}
\alias{.gsd.row.index}
\title{Identify the row number of the geneset being queried.}
\usage{
.gsd.row.index(x, i, j)
}
\arguments{
\item{x}{A GeneSetDb}

\item{i}{the collection of the geneset}

\item{j}{the name of the geneset}
}
\value{
The row index of the geneset under question. If i,j do not match
a given geneset, then NA is returned.
}
\description{
The method only allows querying by single length character vectors
}
