% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{encode_gskey}
\alias{encode_gskey}
\alias{split_gskey}
\title{Converts collection,name combination to key for geneset}
\usage{
encode_gskey(x, y, sep = ";;")

split_gskey(x, sep = ";;")
}
\value{
a character vector

a data.frame with (collection,name) columns
}
\description{
The "key" form often comes out as rownames to matrices and such, or
particularly for sending genesets down into wrapped methods, like do.camera.
}
\examples{
gdf <- exampleGeneSetDF()
gskeys <- encode_gskey(gdf)
gscols <- split_gskey(gskeys)
}
