% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{hasGeneSetCollection}
\alias{hasGeneSetCollection}
\title{Check if a collection exists in the \code{GeneSetDb}}
\usage{
hasGeneSetCollection(x, collection, as.error = FALSE)
}
\arguments{
\item{x}{A \code{GeneSetDb}}

\item{collection}{character vector of name(s) of the collections to query}

\item{as.error}{logical if TRUE, this will error instead of returning FALSE}
}
\value{
logical indicating if this collection exists
}
\description{
Check if a collection exists in the \code{GeneSetDb}
}
