% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.hyperGeometricTest.R
\name{hyperGeometricTest}
\alias{hyperGeometricTest}
\title{Perform Hypergeometric Enrichment tests across GeneSetDb.}
\usage{
hyperGeometricTest(gsd, selected, universe, direction = c("over", "under"))
}
\arguments{
\item{gsd}{The \code{GeneSetDb} object to run tests against}

\item{selected}{The ids of the selected features}

\item{universe}{The ids of the universe}

\item{direction}{Same as direction in \code{GOstats}}
}
\value{
A \code{data.table} of results
}
\description{
Perform Hypergeometric Enrichment tests across GeneSetDb.
}

