% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{incidenceMatrix}
\alias{incidenceMatrix}
\title{Creates a 1/0 matrix to indicate geneset membership to target object.}
\usage{
incidenceMatrix(x, y)
}
\arguments{
\item{x}{A \code{GeneSetDb}}

\item{y}{A target (expression) object \code{x} is (or can be) conformed to}
}
\value{
incidence matrix with nrows = number of genesets and
  ncols == nrows(y)
}
\description{
rows are number of genesetes in \code{x}, columns are number of rows in
\code{y}
}

