% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{invalidMethods}
\alias{invalidMethods}
\title{Returns method names that were not run on a MultiGSEAResult}
\usage{
invalidMethods(x, names, as.error = FALSE)
}
\description{
Returns method names that were not run on a MultiGSEAResult
}

