% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-interactive.R
\name{iplot}
\alias{iplot}
\title{Create interactive GSEA plots}
\usage{
iplot(x, y, j, value = c("logFC", "t"), type = c("density", "boxplot",
  "volcano"), main = NULL, ...)
}
\description{
Create interactive GSEA plots
}

