% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{is.conformed}
\alias{is.conformed}
\title{Check if GeneSetDb has been conformed (optionally to a specific
expression object)}
\usage{
is.conformed(x, to)
}
\arguments{
\item{x}{a \code{GeneSetDb}}

\item{to}{an optional expression object}
}
\value{
logical indicating "conform"-ation status in general, or
  specifically to an expression object \code{to}
}
\description{
Check if GeneSetDb has been conformed (optionally to a specific
expression object)
}

