% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateIndividualLogFC.R
\name{is.logFC.like}
\alias{is.logFC.like}
\title{Checks that a provided table is "similar enough" the the result generated
from calculateIndividualLogFC}
\usage{
is.logFC.like(logFC, x, as.error = FALSE)
}
\arguments{
\item{logFC}{the table to check}

\item{x}{An expression-like object to further test against.}
}
\description{
Checks that a provided table is "similar enough" the the result generated
from calculateIndividualLogFC
}

