% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.of.geneset.lists.to.data.table}
\alias{list.of.geneset.lists.to.data.table}
\title{Helper function to create data.table out of list-of-list genesets}
\usage{
list.of.geneset.lists.to.data.table(x, gene.sets, lookup)
}
\arguments{
\item{x}{An "expression thing"}

\item{gene.sets}{A list of lists of genesets into \code{x}}
}
\description{
Helper function to create data.table out of list-of-list genesets
}

