% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_identifiers.R
\name{load_id_xref}
\alias{load_id_xref}
\title{Load the mouse or human ensembl <-> entrez id maps.}
\usage{
load_id_xref(species, ..., as.dt = FALSE)
}
\arguments{
\item{species}{The name of the species to load the entrez <-> ensembl xref
table for. Currently we only provide tables for human and mouse ... and we
shouldn't even provide those in here, but ...}
}
\value{
a table of ensembl and entrez ids for each species
}
\description{
For convenience, entrez <-> ensembl maps come bundled with the is package.
While there are other sactioned ways to achieve this within the bioconductor
ecosystem, we bundle it here to make our lives easier.
}
\details{
These tables are gnerated by the \code{inst/scripts/genereate-id-maps.R}
}
