% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{loadLocal}
\alias{loadLocal}
\title{Read a local RDS file.}
\usage{
loadLocal(filename)
}
\arguments{
\item{filename}{Path to the file to be read.}
}
\value{
Content of file.
}
\description{
Use the readRDS function to load the given file which should be in RDS
format.
}
