% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{logFC}
\alias{logFC}
\title{Extract the individual fold changes statistics for elements in the
expression object.}
\usage{
logFC(x, .external = TRUE)
}
\arguments{
\item{x}{A \code{MultiGSEAResult}}
}
\value{
The log fold change \code{data.table}
}
\description{
Extract the individual fold changes statistics for elements in the
expression object.
}
\examples{
vm <- exampleExpressionSet(do.voom=TRUE)
gdb <- exampleGeneSetDb()
mg <- multiGSEA(gdb, vm, vm$design, 'tumor', methods=NULL)
lfc <- logFC(mg)
}

