% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_processing.R
\name{mapIDType}
\alias{mapIDType}
\title{Helper function to map only a subset of metabolite IDs}
\usage{
mapIDType(features, keytype = "CHEBI", maptype = "ChEBI", returntype = "HMDB")
}
\arguments{
\item{features}{List of metabolite feature IDs of the pathway.}

\item{keytype}{String specifying the ID format in pathway definition.}

\item{maptype}{String specifying the corresponding ID format in multiGSEA.}

\item{returntype}{String identifying the ID type that should be mapped.}
}
\value{
List of mapped metabolite IDs.
}
\description{
This helper function becomes necessary since there are sometimes multiple ID
formats used in a single pathway definition.
}
