% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-utils.R
\name{meltx}
\alias{meltx}
\title{Melts an Expression Container}
\usage{
meltx(x, pd = NULL, fd = NULL, element = "counts", exform = NULL,
  value.name = "expr")
}
\arguments{
\item{x}{A \code{ExpressionSet}, \code{DGEList}, etc.}

\item{pd}{The "phenodata" columns to keep in melted object. Either the
\code{data.frame}, or the column names of the \code{data.frame} that
will be extracted from the expression object.}

\item{fd}{The "feature data" columns to keep in melted object. Either the
\code{data.frame}, or the column names of the \code{data.frame} that
will be extracted from the expression object.}

\item{element}{The name of "the assay" that holds the expression data}

\item{exform}{A function to call on x to get the (possibly transformed)
data}
}
\description{
Melts an Expression Container
}

