% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolitesMapping.R
\docType{data}
\name{metabolitesMapping}
\alias{metabolitesMapping}
\title{Mapping table for metabolites, derived and compiled from four different data sources.}
\format{A tibble with 7 variables and more then 1 Million metabolites:
\describe{
   \item{DTXCID}{DSSTox structure identifier, character}
   \item{DTXSID}{DSSTox substance identifier, character}
   \item{CAS}{CAS registry number, character}
   \item{CID}{Pubchem compound identifier, character}
   \item{HMDB}{Human Metabolome Database identifier (new format), character}
   \item{ChEBI}{Chemical Entities of Biological Interest identifier, character}
   \item{KEGG}{KEGG Compound identifier, character}
}}
\usage{
data(metabolitesMapping)
}
\description{
Four different sourcex of annotated metabolites, i.e., \code{HMDB},
\code{ChEBI}, \code{CompTox}, and \code{KEGG}, have been retrieved to
compile a comprehensive mapping of available metabolite IDs. ID formats
that are present in the mapping table are: 
DTXCID (Comptox),
DTXSID (Comptox), 
CAS-number,
CID (Pubchem), 
HMDB,
ChEBI,
and KEGG
}
\examples{
data(metabolitesMapping)
}
\keyword{datasets}
