% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-mgGeneSetSummaryByGenes.R
\name{mgGeneSetSummaryByGeneUI}
\alias{mgGeneSetSummaryByGene}
\alias{mgGeneSetSummaryByGeneUI}
\title{Module that displays gene sets related to (by membership) a set of genes.}
\usage{
mgGeneSetSummaryByGeneUI(id, mg = NULL)

mgGeneSetSummaryByGene(input, output, session, mgc, features, method, fdr)
}
\description{
Module that displays gene sets related to (by membership) a set of genes.
}

