% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-mgResultFilter.R
\name{mgResultFilterUI}
\alias{mgResultFilter}
\alias{mgResultFilterUI}
\title{Shiny module to pick GSEA method and fdr params used for display}
\usage{
mgResultFilterUI(id, mg = NULL)

mgResultFilter(input, output, session, mgc)
}
\description{
Shiny module to pick GSEA method and fdr params used for display
}

