% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-mgTableBrowser.R
\name{mgTableBrowserUI}
\alias{mgTableBrowser}
\alias{mgTableBrowserUI}
\title{A module to explore the GSEA statistics generated from a multiGESA Run.}
\usage{
mgTableBrowserUI(id, mg = NULL, server = TRUE)

mgTableBrowser(input, output, session, mgc, server = TRUE)
}
\arguments{
\item{id}{the shiny id of the UI module}

\item{mg}{The \code{MultiGSEAResultContainer} object}

\item{server}{whether or not certain things (options, datatables) be
"server" side or not. This isn't thoroughly pushed throughout everything}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{mgc}{the \code{MultiGSEAResultContainer} object}
}
\description{
A module to explore the GSEA statistics generated from a multiGESA Run.
}

