% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-mgVolcano.R
\name{mgVolcanoUI}
\alias{mgVolcanoUI}
\title{Creates a volcano plot from a source input}
\usage{
mgVolcanoUI(id, x, stats = "dge", xaxis = "logFC", yaxis = "padj",
  idx = "idx", hexbin = TRUE)
}
\arguments{
\item{id}{the shiny id of the output widget}

\item{x}{the object to build a volcano plot from}

\item{stats}{the stats to pull from \code{x} (if necessary) to build the
volcano for.}

\item{x}{the name of the column from the stats table to use on x axis}

\item{y}{the name of the column from the stats table to use on y axis}
}
\description{
Creates a volcano plot from a source input
}

