% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{msg}
\alias{msg}
\title{Utility function to cat a message to stderr (by default)}
\usage{
msg(..., file = stderr())
}
\arguments{
\item{...}{pieces of the message}

\item{file}{where to send the message. Defaults to \code{stderr()}}
}
\description{
Utility function to cat a message to stderr (by default)
}
