% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb-broad.R
\name{getMSigGeneSetDb}
\alias{getMSigGeneSetDb}
\title{Fetches a \code{GeneSetDb} from geneset collections defined in MSigDB.}
\usage{
getMSigGeneSetDb(collection, species = "human", with.kegg = FALSE,
  version = .msigdb.version.current)
}
\arguments{
\item{collection}{character vector specifying the collections you want
(c1, c2, ..., c7, h)}

\item{species}{human or mouse?}

\item{with.kegg}{The Broad distributes the latest versions of the KEGG
genesets as part of the c2 collection. These genesets come with a
restricted license, so by default we do not return them as part of the
GeneSetDb. To include the KEGG gene sets when asking for the c2
collection, set this flag to \code{TRUE}.}

\item{version}{the version of the MSigDB database to use.}
}
\value{
a \code{GeneSetDb} object
}
\description{
This provides versioned genesets from gene set collections defined in
\href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB}. The following
versions are included in this package:

\itemize{
  \item v5.1
  \item v5.2
}

Starting from version 5.1, the \code{GeneSetDb} includes the symbols for
the features in the genesets (in \code{gdb@db}) as well as additional meta
information about the genesets (in \code{geneSets(gdb)}) for "subcategory"
(which may be most useful for GO gene sets, eg. CC vs BP vs MF) and
"organism" which defines the organism in which the geneset was derived
from ("Homo sapiens" or "Mus musculus").
}
