% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{multiGSEA.methods}
\alias{multiGSEA.methods}
\title{Lists the supported GSEA methods by multiGSEA}
\usage{
multiGSEA.methods(names.only = TRUE)
}
\arguments{
\item{names.only}{if \code{TRUE} (default), only returns the names of the methods,
otherwise also returns meta information about the methods, such as the
package it is found in.}
}
\value{
a character vector of GSEA names, or a list of metadata for each
method.
}
\description{
Lists the supported GSEA methods by multiGSEA
}
