% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{multiGSEA}
\alias{multiGSEA}
\title{Performs several GSEA analysis over an experiment}
\usage{
multiGSEA(gsd, x, design = NULL, contrast = NULL, methods = c("camera"),
  outdir = NULL, use.cache = TRUE, cache.inputs = FALSE,
  feature.min.logFC = 1, feature.max.padj = 0.1, trim = 0.1, ...)
}
\arguments{
\item{gsd}{The \code{\link{GeneSetDb}} to run enrichment over.}

\item{x}{An ExpressoinSet-like object}

\item{design}{A design matrix for the study}

\item{contrast}{The contrast of interest to analyze}

\item{methods}{A character vector indicating the GSEA methods you want to
run. This can be any combination of:
\enumerate{
  \item{camera} {A \code{\link[limma]{camera}} analysis}
  \item{roast} {A \code{\link[limma]{roast}} analysis}
  \item{gsd} {A \code{\link[limma]{geneSetTest} analysis}}
  \item{hyperGeometricTest}{
    Tests for enrichment of differentially expressed features in each
    geneset. Thresholds of differential expression for the individual
    features of a geneset are defind by the \code{feature.min.logFC} and
    \code{feature.max.padj} parameters.
  }
}}

\item{outdir}{A character vector indicating the directory to use inorder to
save/cache the results from the analysis. This is required if
\code{use.cache=TRUE} or if you want to save intermediate results as they
are being generated.}

\item{use.cache}{Logical to indicate whether or not to take advantage of the
poor-man's caching the multiGSEA does. See the "Caching" section below for
further details, but in short the function will look for a saved result
for a given \code{method} before computing it.}

\item{cache.inputs}{If \code{TRUE}, the inputs are serialized in
\code{outdir/cache}}

\item{feature.min.logFC}{The minimum logFC required for an individual
feature (not geneset) to be considered differentialy expressed. Used in
conjunction with \code{feature.max.padj} primarily for summarization
of genesets (by \code{\link{geneSetFeatureStatistics}}), but can also be
used by GSEA methods that require differential expression calls at the
individual feature level, like \code{hyperGeometricTest}.}

\item{feature.max.padj}{The maximum adjusted pvalue used to consider an
individual feature (not geneset) to be differentially expressed. Used in
conjunction with \code{feature.min.logFC}.}

\item{trim}{The amount to trim when calculated trimmed \code{t} and
\code{logFC} statistics for each geneset. This is passed down to the
\code{\link{geneSetFeatureStatistics}} function.}

\item{...}{The arguments are passed down into the various geneset analysis
  functions.}
}
\value{
A \code{MultiGSEAResult}
}
\description{
multiGSEA is wrapper function which delegates GSEA analyses to several
different "workers," which are implemented in different packages. Currently
this primarily wraps the GSEA functinos from limma, but I intend to add
\code{npGSEA} as a supported anlysis soon. Refer to the \code{method}
documentation to see which GSEA methods are supported.
}
\details{
This function will write several results to an output directory if
  a valid value for \code{outdir} is provided. Furthermore, the results of
  these analysis will be cached so that time is saved on subsequent calls.
}
\section{Caching}{


  (NOTE: this documentation needs further work, and this feature isn't
  completely functional yet.)

  Running a GSEA can be a costly propositin, especially when using methods
  that rely on permutation, such as (m)roast. When \code{outdir} is set, the
  inputs and intermediate results from each GSEA method are saved in a
  directory tree that is rooted at \code{outdir}. This parameter works
  intimately with \code{use.cache} and \code{plots.generate}.

  While this can be helpful for manually checking results as they are
  generated, you can also use it to speed up future re-analyses of the data
  -- perhaps you want to add one more method to the GSEA comparison that you
  didn't include originally. The result of each analysis method is stored
  in the \code{cache} directory encoded with the contrast that was tested.
  If that method is run again, and \code{use.cache} is set to \code{TRUE},
  then the result will be loaded if it is found instead of recalculating
  the result again.
}
\seealso{
\code{\link[limma]{camera}},
  \code{\link[limma]{roast}},
  \code{\link[limma]{geneSetTest}}
}

