% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MultiGSEAResult-class}
\alias{MultiGSEAResult-class}
\title{A class that holds all of the things you need to analyze the results from
the multiGSEA call.}
\description{
This is a WIP and we aren't using this yet, but will soon.
}
\section{Slots}{

\describe{
\item{\code{gst}}{The \linkS4class{GeneSetTable} this analysis was run over}

\item{\code{results}}{The list of data.tables that each method returns from their
analyses.}

\item{\code{summary}}{The uber cbind'd data.table with results from all methods in
prefixed columns}

\item{\code{logFC}}{The differential expression statistics for each of the elements
measured in the original experiment}

\item{\code{outdir}}{The directory the images/cache/etc. are saved in.}
}}

