% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{geneSetDb}
\alias{geneSetDb}
\title{Fetches the GeneSetDb from MultiGSEAResult}
\usage{
geneSetDb(x)
}
\arguments{
\item{x}{\code{MultiGSEAResult}}
}
\value{
The \code{GeneSetDb}
}
\description{
Fetches the GeneSetDb from MultiGSEAResult
}
\examples{
vm <- exampleExpressionSet(do.voom=TRUE)
gdb <- exampleGeneSetDb()
mg <- multiGSEA(gdb, vm, vm$design, 'tumor', methods=NULL)
geneSetDb(mg)
}
