% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-MultiGSEAResultContainer.R
\name{MultiGSEAResultContainer}
\alias{MultiGSEAResultContainer}
\title{A wrapper for a MultiGSEAResult that makes explicity some shiny bits}
\usage{
MultiGSEAResultContainer(x)
}
\arguments{
\item{x}{A \code{\link[multiGSEA]{MultiGSEAResult}} object, or a path to
an rds-serliazed one}
}
\value{
a \code{MultiGSEAResultContainer} object
}
\description{
A wrapper for a MultiGSEAResult that makes explicity some shiny bits
}

