% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{ncbi.entrez.link}
\alias{ncbi.entrez.link}
\title{Transforms a feature table to have the symbols link to NCBI webpage}
\usage{
ncbi.entrez.link(x, link.col = "symbol")
}
\arguments{
\item{x}{a data.frame from \code{logFC(MultiGSEAResult)}}

\item{link.col}{the column in \code{x} that should be transformed to a link}
}
\value{
a modified \code{x} with an html link in \code{link.col}
}
\description{
Transforms a feature table to have the symbols link to NCBI webpage
}

