% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{org}
\alias{org}
\alias{org<-}
\title{Query/set the organism for a given geneset (db or collection)}
\usage{
org(x, i, ...)

org(x, i) <- value
}
\description{
Query/set the organism for a given geneset (db or collection)
}

