% Generated by roxygen2 (4.0.2): do not edit by hand
\name{p.matrix}
\alias{p.matrix}
\title{Assembles a matrix of nominal or adjusted pvalues from a multiGSEA result}
\usage{
p.matrix(x, pval = c("adjusted", "nominal"))
}
\arguments{
\item{x}{The output from multiGSEA}

\item{pval}{Are we testing pvalues or adjusted pvalues?}
}
\value{
A matrix of all of the
}
\description{
Assembles a matrix of nominal or adjusted pvalues from a multiGSEA result
}

