% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.barcode}
\alias{plot.barcode}
\title{limma's barcodeplot of one or two genesets (with enrichment)}
\usage{
plot.barcode(statistics, index = NULL, index2 = NULL, gene.weights = NULL,
  weights.label = "Weight", labels = c("Up", "Down"), quantiles = c(-1,
  1), col.bars = NULL, worm = TRUE, span.worm = 0.45, ...)
}
\description{
This was imported from the barcode plot in limma_3.22.1 (the function was
annotated with a last update date of 20 October 2008)
}

