% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plots.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,MultiGSEAResult,ANY-method}
\title{Static versions of the iplot functionality}
\usage{
plot(x, y, ...)

\S4method{plot}{MultiGSEAResult,ANY}(x, y, j, value = c("logFC", "t"),
  type = c("density", "barcode", "mini"), main = NULL, ...)
}
\description{
You can create static versions of the iplot functionality (density only).
This also provides a \code{type='barcode'} functionality that draws
limma style barcode plots.

Note that the limma barcode plot is a direct copy of limma's functionality
into here, tweaked ever so slightly so that it's easy to pass GeneSetDb
information down through to it.
}
