% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_category_overlap_heatmap.R
\name{plot_category_overlap_heatmap}
\alias{plot_category_overlap_heatmap}
\title{Plot overlap between categories as a heatmap}
\usage{
plot_category_overlap_heatmap(gene_list_overlaps, minoverlap = 0)
}
\arguments{
\item{gene_list_overlaps}{Output of create_category_overlap_matrix}

\item{minoverlap}{Minimum percentage overlap to plot}
}
\value{
A ggplot heatmap of the overlaps between categories
}
\description{
Plot overlap between categories as a heatmap
}
