% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.enrichtest.R
\name{plot_enrichtest_bias}
\alias{plot_enrichtest_bias}
\title{Plots bias of coviarate to DE / selected status}
\usage{
plot_enrichtest_bias(
  x,
  selected,
  feature.bias,
  title = "DE status vs bias",
  ...
)
}
\description{
The meat and potatoes of this function's code was extracted from
limma::kegga, originally written by Gordon Smyth and Yifang Hu.
}
