% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_values_for_category.R
\name{plot_values_for_category}
\alias{plot_values_for_category}
\title{Plot data for single GO category as a heatmap}
\usage{
plot_values_for_category(
  gsea_result,
  fold_change_data,
  category_id,
  min_value = NULL,
  max_value = NULL
)
}
\arguments{
\item{gsea_result}{A full or filtered multi_gsea result}

\item{fold_change_data}{The original quantitative data from which the gsea result was calculated}

\item{category_id}{The Category ID name for the GO category to plot}
}
\value{
A ggplot heatmap of the genes in the category using the quantitative data
}
\description{
Plot data for single GO category as a heatmap
}
