% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{prepareRenderGseaResultTable}
\alias{prepareRenderGseaResultTable}
\title{Prepares the datatable arguments to use for the gsea.result.table display}
\usage{
prepareRenderGseaResultTable(x, mg)
}
\arguments{
\item{x}{The \code{gsea.result.table} \code{data.table} object}

\item{mg}{The \code{MultiGSEAResult} object}
}
\value{
a list of arguments to use for a \code{do.call, datatable, retlist}
}
\description{
Prepares the datatable arguments to use for the gsea.result.table display
}

