% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{readGeneSymbols}
\alias{readGeneSymbols}
\title{Reads in a semi-annotated genelist (one symbol per line)}
\usage{
readGeneSymbols(fn)
}
\arguments{
\item{fn}{the path to the gene list file}
}
\value{
character vector of gene names.
}
\description{
Often we are given a list of gene names, and the symbols provided are not
the official HGNC ones. In these cases (when small enough) I will replace
the symbol provided by the official one, and leave the submitted symbol
there after a comment character ("#"). This just strips the stuff after
the comment character to provide only the offiical symbols.
}
