% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_rows.R
\name{rename_rows}
\alias{rename_rows}
\title{When you have an object with rownames, and you want to swap them by lookup.}
\usage{
rename_rows(x, rename.rows, ...)
}
\arguments{
\item{x}{an object to whose rows need renaming}

\item{rename.rows}{an objec to help with the renaming.
* A character vector where length(rename.rows) == nrow(x). Every row in
  x should correspond to the renamed value in the same position in
  rename.rows
* If x is a DGEList, SummarizedExperiment, etc. this can be a string.
  In this case, this mus be a column in the fData/rowData/etc. The
  values in that column will be the new
* A two column data.frame. The first column has entries in rownames(x),
  and the second column is the value to rename it to.}
}
\description{
For instance, have a SummarizedExperiment, DGEList, matrix, etc. that have
ensembl id rowname`single cell experiment with ENSGID rownames that you
want to convert to symbols? Now you can.
}
