% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{renderFeatureStatsDataTable}
\alias{renderFeatureStatsDataTable}
\title{Creates a DT::datatable of feature level statistics for use in shiny bits}
\usage{
renderFeatureStatsDataTable(x, features = NULL, digits = 3,
  columns = NULL, feature.link.fn = NULL, order.by = "logFC",
  order.dir = c("desc", "asc"), filter = "none", length.opts = c(10, 25,
  50, 100, 250))
}
\arguments{
\item{x}{A \code{MultiGSEAResult} or \code{data.frame} of feature level
statistics. When \code{x} is a \code{\link{MultiGSEAResult}}, the
\code{\link{logFC}} feature level statistics will be extracted for
display.}

\item{features}{A character vector that specifies the subset of
\code{featureId}'s to display from \code{x}. If \code{NULL} (default),
all of \code{x} will be used.}

\item{digits}{number of digits to round the numeric columns to}

\item{columns}{the columns from \code{x} to use. If \code{missing}, then
only \code{c('symbol', 'featureId', 'logFC', 'pval', 'padj', order.by)}
will be used. If explicitly set to \code{NULL} all columns will be used.}

\item{feature.link.fn}{A funcion that receives the data.frame of statistics
to be rendered and transforms one of its columns into a hyperlink for
further reference. Refer to the \code{\link{ncbi.entrez.link}} function
as an example}
}
\description{
We often want to display an interactive table of feature level statistics
for all features. This function is a convenience wrapper to do that.
}

