% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{renderGeneSetStatsDataTable}
\alias{renderGeneSetStatsDataTable}
\title{Prepares the datatable to show the geneset members and their logFC stats}
\usage{
renderGeneSetStatsDataTable(gstats, name, digits = 3)
}
\value{
a list of arguments that you can \code{do.call, datatable}
}
\description{
Prepares the datatable to show the geneset members and their logFC stats
}

