% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{renderGseaResultTableDataTable}
\alias{renderGseaResultTableDataTable}
\title{Prepares the datatable arguments to use for the gsea.result.table display}
\usage{
renderGseaResultTableDataTable(x, method, mg, digits = 3)
}
\arguments{
\item{x}{The \code{gsea.result.table} \code{data.table} object}

\item{mg}{The \code{MultiGSEAResult} object}
}
\value{
a DT::DataTable
}
\description{
Prepares the datatable arguments to use for the gsea.result.table display
}

