% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{result}
\alias{result}
\title{Extract the result from a given enrichment test from a MultiGSEAResult}
\usage{
result(x, name, stats.only = FALSE, rank.by = c("pval", "logFC", "t"),
  add.suffix = FALSE)
}
\arguments{
\item{x}{MultiGSEAResult}

\item{name}{the names of the results desired}

\item{stats.only}{logical, set to \code{FALSE} if you want to return all
(column-wise) data for each result. By default only the pvalues,
adjusted pvalues, and rank are returned.}

\item{rank.by}{the statistic to use to append a \code{rank} column for the
geneset result}

\item{add.suffix}{If \code{TRUE}, adds \code{.name} as a suffix to the
  columns of the \code{method}-specific statistics returned.}
}
\value{
a data.table with the results from the requested method.
}
\description{
Extract the result from a given enrichment test from a MultiGSEAResult
}

