% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{resultNames}
\alias{resultNames}
\title{Fetch names of GSEA methods that were run from a \code{MultiGSEAResult}}
\usage{
resultNames(x)
}
\arguments{
\item{x}{A MultiGSEAResult}
}
\description{
Fetch names of GSEA methods that were run from a \code{MultiGSEAResult}
}

