% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{resultNames}
\alias{result}
\alias{resultNames}
\alias{results}
\alias{tabulateResults}
\title{Fetch names of GSEA methods that were run from a \code{MultiGSEAResult}}
\usage{
resultNames(x)

result(x, name, stats.only = FALSE, rank.by = c("pval", "t", "logFC"),
  add.suffix = FALSE, .external = TRUE)

results(x, names = resultNames(x), stats.only = TRUE, rank.by = c("pval",
  "logFC", "t"), add.suffix = length(names) > 1L, .external = TRUE)

tabulateResults(x, names = resultNames(x), max.p = 0.3, p.col = c("padj",
  "padj.by.collection", "pval"))
}
\arguments{
\item{x}{A MultiGSEAResult}

\item{name}{the names of the results desired}

\item{stats.only}{logical, set to \code{FALSE} if you want to return all
(column-wise) data for each result. By default only the pvalues,
adjusted pvalues, and rank are returned.}

\item{rank.by}{the statistic to use to append a \code{rank} column for the
geneset result. By default we rank by pvalue calculated by the GSEA
method. You can rank the results based on the trimmed mean of the logFC's
calculated for all of the features in the geneset (\code{"logFC"}), the
trimmed t-statistics of the these features (\code{"t"}), or the "J-G"
statistic of the geneset.}

\item{add.suffix}{If \code{TRUE}, adds \code{.name} as a suffix to the
columns of the \code{method}-specific statistics returned, ie. the
\code{pval} column from the \code{camera} result will be turned to
\code{pval.camera}.}

\item{names}{The results you want to cbind together for the output}

\item{max.p}{The maximum padj value to consider a result significant}

\item{p.col}{use padj or padj.by.collection?}

\item{x}{MultiGSEAResult}

\item{x}{\code{MultiGSEAResult}}

\item{stats.only}{logical, set to \code{FALSE} if you want to return all
(column-wise) data for each result. By default only the pvalues,
adjusted pvalues, and rank are returned.}

\item{add.suffix}{If \code{TRUE}, adds \code{.name} as a suffix to the
columns of the \code{method}-specific statistics returned. This is forced
to \code{TRUE} if \code{length(names) > 1L}}

\item{x}{\code{MultiGSEAResult}}

\item{names}{The results you want to cbind together for the output}
}
\value{
a data.table with the results from the requested method.

a data.table with all of the results. The results will optionally
  be suffixed with the method name that generated them when
  \code{length(names) > 1L}

a data.table that summarizes the significant results per method
  per collection for the GSEA that was run
}
\description{
Fetch names of GSEA methods that were run from a \code{MultiGSEAResult}

TODO: Enable a way for caller to get a subset of the genesets tested,
perhaps this will happen at the "collectoin" level. In this case, it's
not clear if the pval correction should be redone to only account for
the genesets that are returned from.

If the result from more than one method is requested, the column names for
these results will be suffixed with the method name.

Create a summary table to indicate number of significant genesets per
collection, per method
}

