% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{ret.df}
\alias{ret.df}
\title{Returns data.table objects in formats people like.}
\usage{
ret.df(x, .external = TRUE)
}
\description{
The data.table package is heavily used internally in this package, and I
want those objects to (always) be data.table objects while they are passed
around from one function to the next. When data.tables are passed back to
the user, they will be automatically converted into to data.frames, by
default.
}
\details{
Internal multiGSEA functions that return a data.table are always
(1) parameterized with an \code{.external=TRUE} parameter; and
(2) pass the outgoing data.table through this function.

If \code{.external=TRUE}, then the data.table will be transformed on the way
out back to whatever is set for \code{getOption('multiGSEA.df.return')}.

\code{getOption('multiGSEA.df.return')} is automatically set to "data.frame"
when the package is loaded if this option isn't already set in the user's
enviroment (ie. in their \code{.Rprofile}, for instance). The assumption is
that user's are most accustomed to using data.frames, and some of the
idiosyncracies of indexing into a data.table won't trip them up.

Does this extra monkey business make the internal code of multiGSEA a bit
more painful to write? Yes. So ... you're welcome.
}
