% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rmd.multi.GSEA}
\alias{rmd.multi.GSEA}
\title{Runs several GSEA methods over a dataset.}
\usage{
rmd.multi.GSEA(result.name, x, gene.sets, design = NULL, contrast = NULL,
  ...)
}
\arguments{
\item{result.name}{The name/directory used to store the results for this
(unique) analysis.}
}
\description{
Runs several GSEA methods over a dataset.
}

