% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round.dt}
\alias{round.dt}
\title{Rounds every numeric column of a data.table to given precision}
\usage{
\method{round}{dt}(x, digits = 3)
}
\arguments{
\item{x}{a \code{data.table} like thing}

\item{digits}{the number of digits to round to}
}
\value{
the rounded \code{x}
}
\description{
Rounds every numeric column of a data.table to given precision
}

