% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-utils.R
\name{RPKM}
\alias{RPKM}
\title{Modfied version of edgeR RPKM for easy access to gene size}
\usage{
RPKM(x, gene.info = NULL, regularized = FALSE,
  impute.missing.lengths = TRUE, id.col = "entrez.id", size.col = "size",
  ...)
}
\arguments{
\item{x}{expression object}

\item{gene.info}{data.table that has the lengths for the genes in x.
If \code{NULL} this will fetch a default gene size table that I extracted
from the TCGA::COAD datasets in ExpressionPlot}

\item{regularized}{passed down to CPM, but set to \code{FALSE} by default
here.}

\item{id.col}{The name of the column in \code{gene.info} that has the IDs
that match to \code{rownames(x)}}

\item{size.col}{The name of the column in \code{gene.info} that has the
gene lengths to normalize by}
}
\value{
a matrix of RPKM values
}
\description{
Modfied version of edgeR RPKM for easy access to gene size
}

