% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multi_gsea.R
\name{run_multi_gsea}
\alias{run_multi_gsea}
\title{Run multiple GSEAs and collate all results}
\usage{
run_multi_gsea(
  data,
  org = org.Hs.eg.db,
  minGSSize = 10,
  maxGSSize = 50,
  ontology = "BP"
)
}
\arguments{
\item{data}{\itemize{
\item a tibble with a column called Genes containing gene names and all other columns containing quantitative values to use for GSEA
}}

\item{org}{\itemize{
\item the Bioconductor organism to use for the analysis (eg org.Hs.eg.db for human, org.Mm.eg.db for mouse)
}}

\item{minGSSize}{\itemize{
\item The smallest gene set size to use
}}

\item{maxGSSize}{\itemize{
\item The largest gene set size to use
}}
}
\value{
A tibble of combined results from all conditions
}
\description{
Run multiple GSEAs and collate all results
}
