% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{scoreSingleSamples}
\alias{scoreSingleSamples}
\title{Generates single sample gene set scores across a datasets by many methods}
\usage{
scoreSingleSamples(gdb, y, methods = "ssgsea", as.matrix = FALSE,
  drop.sd = 1e-04, verbose = FALSE, ...)
}
\arguments{
\item{gdb}{A GeneSetDb}

\item{y}{An expression matrix to score genesets against}

\item{methods}{A character vector of methods to score samples by}

\item{as.matrix}{Return results as a list of matrices instead of a melted
data.frame? Defaults to \code{FALSE}.}

\item{drop.sd}{Genes with a standard deviation across columns in \code{y}
with a standard deviation less than this value will be dropped.}

\item{verbose}{make some noise? Defaults to \code{FALSE}.}
}
\value{
\code{matrix} with as many rows as \code{geneSets(gdb)} and
  as many columns as \code{ncol(x)}
}

