% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/do.hyperGeometricTest.R
\name{selectedByHyperG}
\alias{selectedByHyperG}
\title{Identify the features that were "selected" for a given geneset for the
hyperGeometricTest for enrichment.}
\usage{
selectedByHyperG(x, i, j)
}
\arguments{
\item{x}{a \code{MultiGSEAResult}}

\item{i}{the collection of the gene set}

\item{j}{the id of the geneset}
}
\value{
A character vector of features in the expression object that were
  identified as part of the geneset that was "selected"
}
\description{
Identify the features that were "selected" for a given geneset for the
hyperGeometricTest for enrichment.
}

