% Generated by roxygen2 (4.0.2): do not edit by hand
\name{significantGeneSets}
\alias{significantGeneSets}
\title{Identifies which geneset has a significant result in >= 1 GSEAs}
\usage{
significantGeneSets(x, pval = c("adjusted", "nominal"), threshold = 0.1,
  n.sig = 1L)
}
\arguments{
\item{x}{The output from multiGSEA}

\item{pval}{Are we testing pvalues or adjusted pvalues?}

\item{threshold}{The threshold that \code{pval} has to be less than}

\item{n.sig}{The number of pval|padj columns that need to hit significance
  for the geneset to be flagged as such}
}
\value{
a logical vector as long as \code{nrow(x)} with \code{TRUE}
  indicates the gene sets that are significant given the criteria
}
\description{
Identifies which geneset has a significant result in >= 1 GSEAs
}

