% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_info}
\alias{species_info}
\title{Match a species query to the regularized species info.}
\usage{
species_info(query, ...)
}
\arguments{
\item{query}{the species name to lookup}
}
\description{
Match a species query to the regularized species info.
}
