% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{ssGSEA.normalize}
\alias{ssGSEA.normalize}
\title{Normalize a vector of ssGSEA scores in the ssGSEA way.}
\usage{
ssGSEA.normalize(x, bounds = range(x))
}
\arguments{
\item{x}{a \code{numeric} vector of ssGSEA scores for a single signature}

\item{bounds}{the maximum and minimum scores obvserved used to normalize
against.}
}
\value{
normalized \code{numeric} vector of \code{x}
}
\description{
ssGSEA normalization (as implemented in GSVA (ssgsea.norm)) normalizes the
individual scores based on ALL scores calculated across samples AND
genesets. It does NOTE normalize the scores within each geneset
independantly of the others.
}

