% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\docType{methods}
\name{[,GeneSetDb,ANY,ANY,ANY-method}
\alias{[,GeneSetDb,ANY,ANY,ANY-method}
\title{Subset whole genesets from a GeneSetDb}
\usage{
\S4method{[}{GeneSetDb,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\description{
Subset whole genesets from a GeneSetDb
}
