% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{subset.MultiGSEAResult}
\alias{subset.MultiGSEAResult}
\title{Subset MultiGSEAResult to include include results for specified genesets}
\usage{
\method{subset}{MultiGSEAResult}(x, keep)
}
\arguments{
\item{x}{MultiGSEAResult}

\item{keep}{logical vector as long as there are numbers of results}
}
\value{
a \code{MultiGSEAResult} that has only the results for the specified
  genesets.
}
\description{
This isn't exported yet because I don't like its implementation
}

