% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{subsetByFeatures}
\alias{subsetByFeatures}
\title{Subset a GeneSetDb to only include geneSets with specified features.}
\usage{
subsetByFeatures(x, featureIds, ...)
}
\arguments{
\item{x}{\code{GeneSetDb}}

\item{featureIds}{Character vector of featureIds}
}
\value{
A subset of \code{x} which contains only the geneSets that contain
  features found in \code{featureIds}
}
\description{
Subset a GeneSetDb to only include geneSets with specified features.
}

