% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{tabulateResults}
\alias{tabulateResults}
\title{Create a summary table to indicate number of significant genesets per
collection, per method}
\usage{
tabulateResults(x, names = resultNames(x), max.p = 0.3, p.col = c("padj",
  "padj.by.collection", "pval"))
}
\arguments{
\item{x}{\code{MultiGSEAResult}}

\item{names}{The results you want to cbind together for the output}

\item{max.p}{The maximum padj value to consider a result significant}

\item{p.col}{use padj or padj.by.collection?}
}
\value{
a data.table that summarizes the significant results per method
  per collection for the GSEA that was run
}
\description{
Create a summary table to indicate number of significant genesets per
collection, per method
}

