% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tricubeMovingAverage}
\alias{tricubeMovingAverage}
\title{Moving average smoother with tricube distance weights to a numeric vector.}
\usage{
tricubeMovingAverage(x, span = 0.5, power = 3)
}
\arguments{
\item{x}{numeric vector}

\item{span}{the smoother span. This gives the proportion of \code{x} values
that contribute to each moving average. Larger values give more
smoothness. Should be positive but not greater than 1.}

\item{power}{a positive exponent used to compute the tricube weights.
\code{power=3} gives the usual tricube weights.  Smaller values give more
even weighting.  Should be greater than 0.}
}
\description{
Copied from limma to sidestep issue of dplyr trampling the call to
\code{filter} here.
}

