% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-class.R
\name{.validateCollectionMetadata}
\alias{.validateCollectionMetadata}
\title{Checks validaty of collectionMetadata of a GeneSetDb}
\usage{
.validateCollectionMetadata(object)
}
\arguments{
\item{object}{A \code{GeneSetDb}}
}
\value{
TRUE if the collectionMetadata is kosher, otherwise a character
  vector of errors.
}
\description{
This function is for internal use only
}
\details{
The following assertions are tested:
\itemize{
  \item{All collection,name entries are unique}
  \item{All collections have a url_function}
  \item{
    The collections that are listed in collectionMetadata have >= 1 defined
    genesets in \code{geneSets(object)}
  }
}
}

