% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Create an interactive volcano plot}
\usage{
volcano_plot(x, stats = "dge", xaxis = "logFC", yaxis = "pval", idx,
  xtfrm = base::identity, ytfrm = function(vals) -log10(vals),
  xlab = xaxis, ylab = sprintf("-log10(\%s)", yaxis), highlight = NULL,
  horiz_lines = c(padj = 0.1), xhex = NULL, yhex = NULL, point.size = 5,
  tools = c("box_select", "reset", "save"), width = NULL, height = NULL,
  shiny_source = "mgvolcano", ggtheme = theme_bw(), ...)
}
\arguments{
\item{highlight}{A vector of featureIds to highlight, or a GeneSetDb
that we can extract the featureIds from for this purpose.}
}
\description{
Create an interactive volcano plot
}
