% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{all.equal.GeneSetDb}
\alias{all.equal.GeneSetDb}
\title{Checks equality (feature parity) between GeneSetDb objects}
\usage{
\method{all.equal}{GeneSetDb}(target, current, features.only = FALSE, ...)
}
\description{
Checks equality (feature parity) between GeneSetDb objects
}

