% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{annotateGeneSetMembership}
\alias{annotateGeneSetMembership}
\title{Annotates rows of a data.frame with geneset membership from a GeneSetDb}
\usage{
annotateGeneSetMembership(x, gdb, x.ids = NULL, ...)
}
\arguments{
\item{x}{A data.frame with genes/features in rows}

\item{gdb}{A \code{GeneSetDb} object with geneset membership}

\item{x.ids}{The name of the column in \code{x} that holds the feautre
id's in \code{x} that match the featureId's in \code{gdb}, or a vector
of id's to use for each row in \code{x} that represent these.}
}
\value{
Returns the original \code{x} with additional columns: each is a
  logical vector that indicates membership for genesets defined in
  \code{gdb}.
}
\description{
This is helpful when you don't have a monsterly sized GeneSetDb. There will
be as many new columns added to \code{x} as there are active genesets in
\code{gdb}.
}
\examples{
vm <- exampleExpressionSet(do.voom=TRUE)
gdb <- getMSigGeneSetDb('h', 'human')
mg <- multiGSEA(gdb, vm, vm$design, 'tumor', methods=NULL)
lfc <- logFC(mg)
annotated <- annotateGeneSetMembership(lfc, gdb, 'featureId')

## Show only genes that are part of 'HALLMARK_ANGIOGENESIS' geneset
(angio <- subset(annotated, `h;;HALLMARK_ANGIOGENESIS`))
}

