% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{approx.target.from.transformed}
\alias{approx.target.from.transformed}
\title{Get the approximate nominal pvalue for a target qvalue given the
distribution of adjusted pvalues from the nominal ones.}
\usage{
approx.target.from.transformed(target, orig, xformed, thresh = 0.01)
}
\arguments{
\item{target}{the FDR value you are trying to find on the nominal pvalue
space (y-axis). This is a value on the FDR scale}

\item{thresh}{how close padj has to be in padjs to get its nominal
counterpart}

\item{pvals}{the distribution of nominal pvalues}

\item{padjs}{the adjusted pvalues from \code{pvals}}
}
\value{
numeric answer, or NA if can't find nominal pvalue within given
  threshold for padjs
}
\description{
This is not an exported function, so you shouldn't be using this. This is
used in the volcano plot code to identify the value on the y-axis of a
nominal pvalues that a given corrected pvalue (FDR) lands at.
}
