% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{as.data.frame.GeneSetDb}
\alias{as.data.frame.GeneSetDb}
\title{Convert a GeneSetDb to a data.frame}
\usage{
\method{as.data.frame}{GeneSetDb}(x, value = c("featureId", "x.id", "x.idx"),
  active.only = is.conformed(x), ...)
}
\arguments{
\item{x}{A \code{GeneSetDb} object}

\item{value}{The value to use for the individual feature id's per collection}

\item{active.only}{If the \code{GeneSetDb} is conformed, do you want to only
return the features that match target and are "active"?}

\item{...}{nothing}
}
\value{
a \code{data.frame} of the GeneSetDb
}
\description{
Convert a GeneSetDb to a data.frame
}

