% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_matrix}
\alias{as_matrix}
\title{Converts an expression container like object to a matrix for analysis}
\usage{
as_matrix(y)
}
\description{
Converts an expression container like object to a matrix for analysis
}

