% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIndividualLogFC.R
\name{calculateIndividualLogFC}
\alias{calculateIndividualLogFC}
\title{Utility function to run limma differential expression analysis}
\usage{
calculateIndividualLogFC(x, design, contrast = ncol(design),
  robust.fit = FALSE, robust.eBayes = FALSE, trend.eBayes = FALSE,
  treat.lfc = NULL, confint = TRUE, with.fit = FALSE, ...,
  as.dt = FALSE)
}
\arguments{
\item{x}{The expression object. This can be 1 column matrix if you are not
running any analysis, and this function essentially is just a "pass through"}

\item{design}{The design matrix for the experiment}

\item{contrast}{The contrast you want to test and provide stats for}

\item{robust.fit}{If \code{TRUE}, sets \code{method='robust'} in the
\code{lmFit} call. Defaults to \code{FALSE}}

\item{robust.eBayes}{The value of the \code{robust} parameter in the
\code{eBayes} call.}

\item{treat.lfc}{If this is numeric, this activates limma's "treat"
functionality and tests for differential expression against this
specified log fold change threshold. This defaults to \code{NULL}.}

\item{confint}{add confidence intervals to \code{topTable} output (default
\code{TRUE})? Ignored if \code{x} is a \code{DGEList}.}

\item{with.fit}{If \code{TRUE}, this function returns the fit in addition
to the logFC statistics.}

\item{...}{parameters passed through to the \code{lmFit} call.}

\item{as.dt}{Return the result as a \code{data.table}? Defaults to
\code{FALSE}.}
}
\value{
If \code{with.fit == FALSE} (the default) a \code{data.table} of
  logFC statistics for the contrast under test. Otherwise, a list is
  returned with \code{$result} containing the logFC statistics, and
  \code{$fit} has the limma fit for the data/design/contrast under test.
}
\description{
Utility function to run limma differential expression analysis
}
\details{
If \code{x} is a \code{DGEList} we require that \code{estimateDisp} has
already been called and this will run a quasi-likelihood based differential
expression analysis, otherwise a "vanilla" limma run will be run.

Lastly, if \code{x} is simply a single column matrix, we assume that we are
just passing a single pre-ranked vector of statistics through multiGSEA's
analysis pipelines, and a logFC-like data.frame is constructed with these
statistics in the \code{logFC} and \code{t} columns.
}
