% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculateIndividualLogFC.R
\name{calculateIndividualLogFC}
\alias{calculateIndividualLogFC}
\title{Runs a vanilla limma differential expression analysis}
\usage{
calculateIndividualLogFC(x, design, contrast = ncol(design),
  robust.fit = FALSE, robust.eBayes = FALSE, with.fit = FALSE, ...)
}
\arguments{
\item{x}{The expression object. This can be 1 column matrix if you are not
running any analysis, and this function essentially is just a "pass through"}

\item{design}{The design matrix for the experiment}

\item{contrast}{The contrast you want to test and provide stats for}

\item{robust.fit}{If \code{TRUE}, sets \code{method='robust'} in the
\code{lmFit} call. Defaults to \code{FALSE}}

\item{robust.eBayes}{The value of the \code{robust} parameter in the
\code{eBayes} call.}

\item{...}{parameters passed through to the \code{lmFit} call.}

\item{provide}{Specifies what you want to get out of this, just the
\code{logFC} or \code{t} statistics, or the whole result \code{table}.
TODO: remove this parameter and just always return the table, if one is
generated as a result of the function call (ie. \code{ncol(x) > 1})}
}
\value{
A data.table with the logFC results per row of \code{x}. The
  \code{$featureId} column is the rownames (ids) of x.
}
\description{
Runs a vanilla limma differential expression analysis
}

