% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIndividualLogFC.R
\name{calculateIndividualLogFC}
\alias{calculateIndividualLogFC}
\title{Runs a vanilla limma differential expression analysis}
\usage{
calculateIndividualLogFC(x, design, contrast = ncol(design),
  robust.fit = FALSE, robust.eBayes = FALSE, trend.eBayes = FALSE,
  use.treat = FALSE, treat.lfc = log2(1.25), confint = TRUE,
  with.fit = FALSE, ..., .external = TRUE)
}
\arguments{
\item{x}{The expression object. This can be 1 column matrix if you are not
running any analysis, and this function essentially is just a "pass through"}

\item{design}{The design matrix for the experiment}

\item{contrast}{The contrast you want to test and provide stats for}

\item{robust.fit}{If \code{TRUE}, sets \code{method='robust'} in the
\code{lmFit} call. Defaults to \code{FALSE}}

\item{robust.eBayes}{The value of the \code{robust} parameter in the
\code{eBayes} call.}

\item{use.treat}{logical indicating whether or not to use the "treat"
functionality to calculate pvalues based on minimum logFC
(\code{treat.lfc})}

\item{treat.lfc}{The logFC to test the dge on.}

\item{confint}{add confidence intervals to \code{topTable} output (default
\code{TRUE})? Ignored if \code{x} is a \code{DGEList}.}

\item{with.fit}{If \code{TRUE}, this function returns the fit in addition
to the logFC statistics.}

\item{...}{parameters passed through to the \code{lmFit} call.}

\item{.external}{If called w/in multiGSEA package, we don't check to convert
outgoing data.table to user's preferred type. \code{.external} is by
default \code{TRUE}, as if someone on the outside is calling.}
}
\value{
If \code{with.fit == FALSE} (the default) a \code{data.table} of
  logFC statistics for the contrast under test. Otherwise, a list is
  returned with \code{$result} containing the logFC statistics, and
  \code{$fit} has the limma fit for the data/design/contrast under test.
}
\description{
Note that differential expression analysis is always run through limma,
so if \code{x} is a \code{DGEList}, it will be voomd first and then
processed "as usual".
}
\details{
importFrom edgeR glmQLFit glmTreat glmQLFTest topTags
}
