% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateIndividualLogFC}
\alias{calculateIndividualLogFC}
\title{Runs a vanilla limma differential expression analysis}
\usage{
calculateIndividualLogFC(x, design, contrast, provide = c("table", "logFC",
  "t"), robust.fit = FALSE, robust.eBayes = FALSE, with.fit = FALSE, ...)
}
\arguments{
\item{x}{The expression object. This can be 1 column matrix if you are not
running any analysis, and this function essentially is just a "pass through"}

\item{design}{The design matrix for the experiment}

\item{contrast}{The contrast you want to test and provide stats for}

\item{provide}{Specifies what you want to get out of this, just the
\code{logFC} or \code{t} statistics, or the whole result \code{table}.
TODO: remove this parameter and just always return the table, if one is
generated as a result of the function call (ie. \code{ncol(x) > 1})}

\item{robust.fit}{If \code{TRUE}, sets \code{method='robust'} in the
\code{lmFit} call. Defaults to \code{FALSE}}

\item{robust.eBayes}{The value of the \code{robust} parameter in the
\code{eBayes} call.}

\item{...}{parameters passed through to the \code{lmFit} call.}
}
\value{
If \code{provde == 'table'}: a data.frame of results that resembles
  the output of limma::topTable with P.Value and adj.P.Val columns renamed
  to pval, padj. Even if the \code{x} input is a vector, we reconstruct a
  "dummy" \code{data.frame}. Otherwise, the column vector that corresponds
  to \code{provide}.
}
\description{
Runs a vanilla limma differential expression analysis
}

