% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneSetDb-methods.R
\docType{methods}
\name{conform}
\alias{conform}
\alias{conform,GeneSetDb-method}
\title{Map featureId's in a GeneSetDb to a target expression object.}
\usage{
conform(x, ...)

\S4method{conform}{GeneSetDb}(x, y, unique.by = c("none", "mean", "var"),
  min.gs.size = 3L, max.gs.size = Inf, match.tolerance = 0.25)
}
\arguments{
\item{x}{The GeneSetDb}

\item{y}{The expression object/matrix to conform to}

\item{unique.by}{If there are multiple rows that map to the identifiers
used in the genesets, this is a means to pick the single row for that ID}

\item{min.gs.size}{Ensure that the genesets that make their way to the
\code{GeneSetDb@table} are of a minimum size}

\item{max.gs.size}{Ensure that the genesets that make their way to the
  \code{GeneSetDb@table} are smaller than this size}
}
\value{
A \code{GeneSetDb} that has been matched/conformed to an
  expression object target (\code{y})
}
\description{
Map featureId's in a GeneSetDb to a target expression object.

This uses the already-set featureIdMap for the GeneSetDb
}

