% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{constructGseaResultTable}
\alias{constructGseaResultTable}
\title{Builds the subset of the GSEA statistics to present to the user}
\usage{
constructGseaResultTable(mg, method, fdr)
}
\arguments{
\item{mg}{\code{MultiGSEAResult} object}

\item{method}{the method to show statistics for}

\item{the}{FDR cut off to present statistics for}
}
\value{
a data.table of the statistics that match the filtering criteria.
  A 0-row data.table is returned if nothing passes.
}
\description{
This function sets the collection to a factor and puts the MSigDB Hallmark
collection ("h") at the top of the totem pole, if it is included in
\code{mg}.
}

