% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-utils.R
\name{CPM2TPM}
\alias{CPM2TPM}
\title{Convert CPM matrix to TPM}
\usage{
CPM2TPM(x, gene.info = NULL, id.col = "entrez.id", size.col = "size",
  impute.missing.lengths = TRUE, ...)
}
\arguments{
\item{x}{An expression matrix of counts per million (on the natural scale)}
}
\value{
A matrix of TPMs
}
\description{
Convert CPM matrix to TPM
}

