% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.camera.R
\name{do.camera}
\alias{do.camera}
\title{Runs camera on the given experiment.}
\usage{
do.camera(gsd, x, design, contrast = ncol(design), outdir = NULL,
  use.cache = TRUE, weights = NULL, use.ranks = FALSE,
  allow.neg.cor = TRUE, trend.var = FALSE, sort = FALSE, ...)
}
\description{
Runs camera on the given experiment.
}

