% Generated by roxygen2 (4.0.2): do not edit by hand
\name{do.camera}
\alias{do.camera}
\title{Runs camera on the given experiment.}
\usage{
do.camera(x, gs.table, design, contrast, outdir = NULL, use.cache = TRUE,
  weights = NULL, use.ranks = FALSE, allow.neg.cor = TRUE,
  trend.var = FALSE, sort = FALSE, ...)
}
\description{
Runs camera on the given experiment.
}

