% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.camera.R
\name{do.camera}
\alias{do.camera}
\title{Runs camera on the given experiment.}
\usage{
do.camera(gsd, x, design, contrast = ncol(design), ...)
}
\description{
Runs camera on the given experiment.
}
\section{Gene Set Enrichment Methods}{


camera tends to be very conservative, especially when trying to estimate
residual correlation of gene sets from experiments with small N. For this
reason,  Gordon has added the ability provide a prespecified
\code{inter.gene.correlation} value.
\href{https://support.bioconductor.org/p/70005/#70195}{
He suggests to try a small positive number (0.05)}
}

