% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.fgsea.R
\name{do.fgsea}
\alias{do.fgsea}
\title{Runs GSEA on a pre-ranked list of differential expression statistcis with fgsea}
\usage{
\method{do}{fgsea}(gsd, x, design, contrast = ncol(design), minSize = 15,
  maxSize = 15, nperm = 10000, gseaParam = 1, score.by = c("t", "logFC",
  "pval"), use.treat = FALSE, feature.min.logFC = if (use.treat) log2(1.25)
  else 1, feature.max.padj = 0.1, gs.idxs = as.list(gsd, active.only = TRUE,
  value = "x.idx"), logFC = NULL, ...)
}
\arguments{
\item{gsd}{The \code{\link{GeneSetDb}} for analysis}

\item{x}{The expression object. This can be 1 column matrix if you are not
running any analysis, and this function essentially is just a "pass through"}

\item{design}{The design matrix for the experiment}

\item{contrast}{The contrast you want to test and provide stats for}

\item{...}{arguments to pass down into \code{calculateIndividualLogFC}}
}
\value{
A data.table of fgsea results.
}
\description{
Techinically this doesn't require a full design to run, but it proves
expedient for a quick turnaround.
}
\details{
Note that fgsea isn't added to import or suggests because rescomp's
compiler can't handle this yet.

fgsea expects the pathways to be describes as a list of character vectors
where the vectors are gene ids, and the names of the pre-ranked vector
correspond to those IDs
}
