% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.geneSetTest.R
\name{do.geneSetTest}
\alias{do.geneSetTest}
\title{Worker function to run geneSetTest from within a multiGSEA pipeline}
\usage{
do.geneSetTest(gsd, x, design, contrast, outdir = NULL, robust.fit = FALSE,
  robust.eBayes = FALSE, logFC = NULL, score.by = c("t", "logFC", "pval"),
  gs.idxs = as.list(gsd, active.only = TRUE, value = "x.idx"), ...)
}
\description{
\strong{This function is not meant to be called directly, it should only be
called internally within \code{multiGSEA}}
}
