% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.hyperGeometricTest.R
\name{do.hyperGeometricTest}
\alias{do.hyperGeometricTest}
\title{Performs a hypergeometric test for significance of gene set membership.}
\usage{
do.hyperGeometricTest(gsd, x, design, contrast = ncol(design),
  direction = c("over", "under"), logFC = NULL, feature.min.logFC = 1,
  feature.max.padj = 0.1, vm = x, ...)
}
\arguments{
\item{gsd}{The \code{\link{GeneSetDb}} for analysis}

\item{x}{The expression object}

\item{design}{Experimental design}

\item{contrast}{The contrast to test}

\item{direction}{Same as direction in \code{GOstats}}

\item{logFC}{The logFC data.table from \code{calculateIndividualLogFC}}

\item{...}{arguments to pass down into \code{calculateIndividualLogFC}}
}
\description{
Genes are selected for testing against each geneset by virture of them
passing a maximum FDR and minimum log fold change as perscribed by the
\code{min.logFC} and \code{max.padj} parameters, respectfully.
}
\details{
Note that we are intentionally adding a hyperG.selected column by reference
so that this information is kicked back to the caller multiGSEA function
and included in downstream reporting.
}

