% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.logFC.R
\name{do.logFC}
\alias{do.logFC}
\title{Worker function that is never really called within a multiGSEA pipeline}
\usage{
do.logFC(gsd, x, design, contrast = ncol(design), outdir = NULL,
  use.cache = TRUE, logFC = NULL, robust.fit = FALSE,
  robust.eBayes = FALSE, ...)
}
\description{
This is a "pass through" function. The caller asks for \code{method="logFC"}
if she only wants to calculate the geneSet statistics for each of the
geneSets, ie. mean.logFC, mean.t for each geneset.

\strong{This function is not meant to be called directly, it should only be
called internally within \code{multiGSEA}}
}
