% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{do.scoreSingleSamples.gsdecon}
\alias{do.scoreSingleSamples.gsdecon}
\title{Jason's method}
\usage{
do.scoreSingleSamples.gsdecon(gdb, y, melted = FALSE, ...)
}
\description{
Jason's method
}

