% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{do.scoreSingleSamples.svd}
\alias{do.scoreSingleSamples.svd}
\title{A no dependency call to GSDecon's eigengene scoring}
\usage{
do.scoreSingleSamples.svd(gdb, y, melted = FALSE, center = TRUE,
  scale = TRUE, uncenter = center, unscale = scale, ...)
}
\description{
TODO: Need to debug this. For some reason the results aren't the same
as what GSDecon methos provides, even though the codepath should be
identical
}

