% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{do.scoreSingleSamples.svd}
\alias{do.scoreSingleSamples.svd}
\title{A simpler call to Jason's eigengene-like scoring}
\usage{
do.scoreSingleSamples.svd(gdb, y, melted = FALSE, uncenter = TRUE,
  unscale = TRUE, ...)
}
\description{
A simpler call to Jason's eigengene-like scoring
}

