% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.svdGeneSetTest.R
\name{do.svdGeneSetTest}
\alias{do.svdGeneSetTest}
\title{Worker function summarizes gene set activity by SVD score and t-test across
contrast}
\usage{
\method{do}{svdGeneSetTest}(gsd, x, design, contrast = ncol(design),
  gs.idxs = as.list(gsd, active.only = TRUE, value = "x.idx"),
  trend.eBayes = FALSE, ...)
}
\description{
The idea is to summarize gene set level activity per sample by their
svd (GSDecon) score to create a geneset x sample matrix. Then uses this
matrix in a limma based framework to assess their "differential expression"

This method seems like it should produce sane results, but I haven't really
tested it yet and is therefore experimental. The fact that the
"svdGeneSetTest" method isn't mentioned anywhere in the documentation yet is
by design. Still, you have found that it exists and you can try to use it
if you like (or, please contact me to tell me why it's a bad idea!)

\strong{This function is not meant to be called directly, it should only be
called internally within \code{multiGSEA}}
}
