% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{exampleExpressionSet}
\alias{exampleExpressionSet}
\title{Fetch a pre-canned expression dataset that can be used for testing and
examples.}
\usage{
exampleExpressionSet(dataset = c("tumor-vs-normal", "tumor-subtype"),
  do.voom = TRUE)
}
\arguments{
\item{dataset}{Character vector indicating what samples wanted, either
\code{"tumor-vs-normal"} for a tumor vs normal dataset from TCGA, or just the tumor samples
from the same (there are different subtypes in there)}

\item{do.voom}{If TRUE, a voomed EList is returned, otherwise an
ExpressionSet of counts.}
}
\description{
The samples in this dataset are a subset of samples from TCGA-BRCA. Note
that this function will load either \code{limma} or \code{Biobase} into
the calling environment so that the user can manipulate the object that is
returned.
}
\details{
Note that calling this functino currently loads/attaches Biobase. This
shouldn't be, but I am tired of trying to debug this one, so am kicking
the can down the road ...
}

