% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{exampleGeneSets}
\alias{exampleGeneSets}
\title{Get sample gene sets}
\usage{
exampleGeneSets(x, unlist = !missing(x))
}
\arguments{
\item{x}{If provided, an expression/matrix object so that the genesets are
returned as (integer) index vectors into the rows of x whose rownames
match the ids in the geneset.}
}
\value{
A list of lists of entrezIDs when \code{as == 'lol'}, or
  a list of integers into the rows of \code{exampleExpressionSet}
  for the genes in the given geneset.
}
\description{
Get sample gene sets
}

