% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{explore}
\alias{explore}
\title{Interactively explore a MultiGSEAResult via a shiny app}
\usage{
explore(x)
}
\arguments{
\item{x}{A \code{MultiGSEAResult} object}
}
\description{
This will launch a shiny application that enables exploratory analysis of
the results from the different GSEA methods run within a
\code{MultiGSEAResult}.

Reference the "shiny-multiGSEA" vignette for more detailed documentation of
the functionality provided by this application.
}
\examples{
\dontrun{
vm <- exampleExpressionSet()
gdb <- exampleGeneSetDb()
mg <- multiGSEA(gdb, vm, vm$design, methods=c('camera', 'fry'))
explore(mg)
}
}

