% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{explore}
\alias{explore}
\title{Invokes a shiny app to interactively explore a MultiGSEAResult}
\usage{
explore(x)
}
\arguments{
\item{x}{A \code{MultiGSEAResult} object}
}
\description{
I hope you're sitting down
}
\examples{
\dontrun{
vm <- exampleExpressionSet()
gdb <- exampleGeneSetDb()
mg <- multiGSEA(gdb, vm, vm$design, methods=c('camera', 'fry'))
explore(mg)
}
}

