% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{failWith}
\alias{failWith}
\title{Utility function to try and fail with grace.}
\usage{
failWith(default = NULL, expr, frame = parent.frame(), silent = FALSE,
  file = stderr())
}
\arguments{
\item{default}{the value to return if \code{expr} fails}

\item{expr}{the expression to take a shot at}

\item{frame}{the frame to evaluate the expression in}

\item{silent}{if \code{TRUE}, sends the error message to \code{msg}}

\item{file}{where msg sends the message}
}
\value{
the result of \code{expr} if successful, otherwis \code{default}
}
\description{
Utility function to try and fail with grace.
}

