% Generated by roxygen2 (4.0.2): do not edit by hand
\name{featureIds}
\alias{featureIds}
\title{Returns the featureIds associated with a given GeneSetTable or one of the
individual gene sets within it.}
\usage{
featureIds(object, ...)
}
\description{
Returns the featureIds associated with a given GeneSetTable or one of the
individual gene sets within it.
}

