% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneSetDb-methods.R
\docType{methods}
\name{featureIds}
\alias{featureIds}
\alias{featureIds,GeneSetDb-method}
\title{Returns the relevant featureIds for a given geneset.}
\usage{
featureIds(x, ...)

\S4method{featureIds}{GeneSetDb}(x, i, j, value, fetch.all = FALSE,
  active.only = is.conformed(x), ...)
}
\arguments{
\item{x}{The GeneSetDb}

\item{i}{The collection level identifier for the geneset}

\item{j}{The name level identifier for the genest}

\item{value}{What form do you want the id's in?
\describe{
  \item{featureId}{the IDs used in the original geneset definitions}
  \item{x.id}{the ids of the features as they are used in the expression
        object}
  \item{x.idx}{The integer index into the expresion object \code{x} that
        the GeneSetDb has been conformed to.}
}}

\item{fetch.all}{By default, this function only returns the IDs for the
geneset that are matched in the target expression object if this
\code{GeneSetDb} has been "conformed" (ie. \code{is.conformed(x) == TRUE}.
Set this to \code{TRUE} if you want to get all featureIds irrespective of
this constraint.}
}
\value{
A vector of identifiers (or indexes into an expression object,
  depending on the \code{value} argument) for the features in the specified
  geneset. NA is returned if the geneset is not "active" (ie. listed in
  geneSets(x))
}
\description{
Gene sets are defined by the unique compound key consisting of their
\code{collection} and \code{name}. To fetch the featureIds associated with
a specific geneset, you must provide values for \code{i} and \code{j}. If
these are missing, then a character vector of all the unique feature ids
within \code{x} are returned.

If the GeneSetDb \code{x} has been conformed to an expression object this
will default to return the featureId's as they are used/matched to the
expression object, otherwise it will return the featureIds used in the
definition of the gene set database.

\code{x} can be either a \code{GeneSetDb} or a \code{MultiGSEAResult}. If
its the latter, then this call simply delegates to the internal
\code{GeneSetDb}.

Fetch the IDs for the a given gene set.
}

