% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneSetDb-methods.R
\docType{methods}
\name{featureIds}
\alias{featureIds}
\alias{featureIds,GeneSetDb-method}
\title{Returns the featureIds associated with a given GeneSetDb or one of the
individual gene sets within it.}
\usage{
featureIds(x, ...)

\S4method{featureIds}{GeneSetDb}(x, i, j, value, fetch.all = FALSE,
  active.only = is.conformed(x), ...)
}
\arguments{
\item{x}{The GeneSetDb}

\item{i}{The collection level identifier for the geneset}

\item{j}{The name level identifier for the genest}

\item{value}{What form do you want the id's in?
\describe{
  \item{featureId}{the IDs used in the original geneset definitions}
  \item{x.id}{the ids of the features as they are used in the expression
        object}
  \item{x.idx}{The integer index into the expresion object \code{x} that
        the GeneSetDb has been conformed to.}
}}

\item{fetch.all}{By default, this function only returns the IDs for the
  geneset that are matched in the target expression object if this
  \code{GeneSetDb} has been "conformed" (ie. \code{is.conformed(x) == TRUE}.
  Set this to \code{TRUE} if you want to get all featureIds irrespective of
  this constraint.}
}
\value{
A vector of identifiers (or indexes into an expression object) for
  the features in the given geneset. NA is returned if the geneset is not
  "active" (ie. listed in geneSets(x))
}
\description{
Returns the featureIds associated with a given GeneSetDb or one of the
individual gene sets within it.

If the GeneSetDb \code{x} has been conformed to an expression object this
will default to return the featureId's as they are used/matched to the
expression object, otherwise it will return the featureIds used in the
definition of the gene set database.
}

