% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{generate.GSEA.plots}
\alias{generate.GSEA.plots}
\title{Generates GSEA plots for the genesets over a given experiment.}
\usage{
generate.GSEA.plots(x, design, contrast, result, outdir, use.cache = TRUE,
  logFC.stats = NULL, padj.threshold = 1, score.by = c("logFC", "t"), ...)
}
\arguments{
\item{x}{The expression object used in the multiGSEA call}

\item{design}{The desing matrix for this GSEA}

\item{contrast}{The contrast being analyzed in the GSEA}

\item{result}{The result data.table returned (or being generated) in the
\code{multiGSEA} call.}

\item{outdir}{The path to the parent directory that holds the output for
this run.}

\item{use.cache}{If true, the function checks if the image already exists
before replotting it.}

\item{padj.threshold}{plot generation to only include genesets that
have some value in an FDR column that is less than or equal to this
threshold.}

\item{logFC}{pre-computed logFCs for the contrast we are running GSEA
against.}
}
\value{
A data.table that has the collection, id, gene set name, image page
  and whether or not the image was plotted columns for each row in \code{x}
}
\description{
Generates GSEA plots for the genesets over a given experiment.
}

