% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-geneSetContrastView.R
\name{geneSetContrastViewUI}
\alias{geneSetContrastView}
\alias{geneSetContrastViewUI}
\alias{updateGeneSetContrastViewGeneSet}
\title{A module to encapsulate browsing differential statistics of a geneset.}
\usage{
geneSetContrastViewUI(id, height = "590px", width = "400px")

geneSetContrastView(input, output, session, mgc, server = TRUE,
  maxOptions = Inf, sep = "_::_")

updateGeneSetContrastViewGeneSet(session, id, label = NULL, choices = NULL,
  selected = NULL, options = list(), server = FALSE)
}
\arguments{
\item{id}{the shiny id of the module}

\item{height}{the height of the module}

\item{server}{are the tabular bits rendered on teh server side}

\item{mg}{the \code{MultiGSEAResultContainer}}
}
\value{
a tagList of html stuff to dump into the UI
}
\description{
The module is meant to be displayed in "a box". A geneset picker is at the
top of the box. The main content consists of a a
\code{\link[miniUI]{miniTabStripPanel}}, which provides a vizual view as
well as a tabular view of the statistics for a particular geneset from a
\code{MultiGSEAResultContainer}
}

