% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{as.expression.indexes}
\alias{as.expression.indexes}
\alias{as.list.GeneSetDb}
\title{Unrolls the GeneSetDb into a list of index vectors per "active" gene set}
\usage{
\method{as.expression}{indexes}(x, value = c("x.idx", "x.id"),
  active.only = is.conformed(x))

\method{as.list}{GeneSetDb}(x, value = c("x.id", "x.idx"),
  active.only = is.conformed(x), ...)
}
\arguments{
\item{x}{A GeneSetDb}

\item{value}{\code{x.idx} returns indexes into the conformed expression
object as integers (ie. row index numbers) and \code{x.id} returns them
as their featureId's as used in the target expression object.}
}
\description{
Unrolls the GeneSetDb into a list of index vectors per "active" gene set

\code{as.list.GeneSetDb} and \code{as.expression.indexes} intentionally
have different default values for the \code{active.only} parameter.
\code{as.list} returns the ids of features, and \code{as.expression.indexes}
returns the integer index into the expression matrix that the
\code{GeneSetDb} is conformed to.
}

