% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneSetDb-class.R
\name{GeneSetDb}
\alias{GeneSetDb}
\title{Create a GeneSetDb to match an input expression object.}
\usage{
GeneSetDb(x, featureIdMap = NULL)
}
\arguments{
\item{x}{A list (of lists) of character vectors that specify
the features in a particular geneset. The top level of the list (of lists)
are the collection ids. The names of the inner lists define the id for the
particular geneset.}

\item{featureIdMap}{A data.frame with  2 character columns. The first
column is the ids of the genes (features) used to identify the genes in
\code{gene.sets}, the second second column are IDs that this should be
mapped to. Useful for testing probelevel microarray data to gene level
geneset information.}

\item{x}{The expression values measured. Could be a matrix, ExpressionSet,
etc. rownames are required on this "thing" for it to work.}

\item{min.gs.size}{The minimum number of features in a geneset required
for testing (this is calculated after gene.sets to expression feature
mapping).}

\item{max.gs.size}{Same as above, but is upper limit of geneset size for
testing.}
}
\description{
Create a GeneSetDb to match an input expression object.
}
\section{Easily mapping rownames(x) to IDs used in gene.sets}{


It may be that the IDs used in the provided \code{gene.sets} are different
than the ones used in \code{rownames(x)}, for instance the IDs in
\code{gene.sets} might (are likely to) be Entrez IDs, and the rownames of
\code{x} might by affy probe set IDs, what then?

This is where the \code{mapping} parameter becomes useful.

In all cases, the \code{featureIds} stored in the \code{GeneSetDb} are
"in the space" of the expression object \code{x} the GeneSetDb has been
"conformed" to. To recover the original IDs used, you could reference the
\code{featureIdMap} table.

TODO: document the GeneSetDb ID mapping more thoroughly.
}

