% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneSetDb-class.R
\name{GeneSetDb}
\alias{GeneSetDb}
\title{Create a GeneSetDb to match an input expression object.}
\usage{
GeneSetDb(x, featureIdMap = NULL, collectionName = NULL)
}
\arguments{
\item{x}{A \code{GeneSetCollection}, a "two deep" list of either
\code{GeneSetCollection}s or lists of character vectors, which are
the gene identifers. The "two deep" list represents the different
collections (top level) at the top level, and each such list is a named
list itself, which represents the gene sets in the given collection.}

\item{featureIdMap}{A data.frame with  2 character columns. The first
column is the ids of the genes (features) used to identify the genes in
\code{gene.sets}, the second second column are IDs that this should be
mapped to. Useful for testing probelevel microarray data to gene level
geneset information.}

\item{collectionName}{If \code{x} represents a singular collection, ie.
a single \code{GeneSetCollection} or a "one deep" (named (by geneset))
list of genesets, then this parameter provides the name for the
collection. If \code{x} is multiple collections, this can be character
vector of same length with the names. In all cases, if a collection name
can't be defined from this, then collections will be named anonymously.
If a value is passed here, it will overide any names stored in the list of
\code{x}.}
}
\description{
Create a GeneSetDb to match an input expression object.
}
\section{Easily mapping rownames(x) to IDs used in gene.sets}{


It may be that the IDs used in the provided \code{gene.sets} are different
than the ones used in \code{rownames(x)}, for instance the IDs in
\code{gene.sets} might (are likely to) be Entrez IDs, and the rownames of
\code{x} might by affy probe set IDs, what then?

This is where the \code{mapping} parameter becomes useful.

In all cases, the \code{featureIds} stored in the \code{GeneSetDb} are
"in the space" of the expression object \code{x} the GeneSetDb has been
"conformed" to. To recover the original IDs used, you could reference the
\code{featureIdMap} table.

TODO: document the GeneSetDb ID mapping more thoroughly.
}

