% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{geneSetFeatureStats}
\alias{geneSetFeatureStats}
\title{Fetch the logFC statistics for a geneset}
\usage{
geneSetFeatureStats(x, i, j, .external = TRUE)
}
\arguments{
\item{x}{\code{MultiGSEAResult}}

\item{i}{The collection name}

\item{j}{The gene set name}
}
\value{
data.table with the stats for the features in the geneset
}
\description{
Fetch the logFC statistics for a geneset
}

