% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneSetDb-methods.R
\docType{methods}
\name{geneSets}
\alias{geneSets}
\alias{geneSets,GeneSetDb-method}
\title{Fetch the active (or all) gene sets from a GeneSetDb or MultiGSEAResult}
\usage{
geneSets(x, ...)

\S4method{geneSets}{GeneSetDb}(x, active.only = is.conformed(x), ...,
  .external = TRUE)
}
\arguments{
\item{x}{Object to retrieve the gene set from, either a \code{GeneSetDb} or
a \code{MultiGSEAResult}.}

\item{active.only}{only look for gene sets that are "active"? Defaults to
\code{TRUE} if \code{x} is conformed to a target expression object, else
\code{FALSE}. See \code{\link{conform}} for further details.}

\item{.external}{If \code{TRUE} (default), honors the user's setting to
return either a \code{data.frame} or \code{data.table} object. This
setting is defined in the \code{"multiGSEA.df.return"} option
(\code{getOption("multiGSEA.df.return"))}). The value for this option
can be either of \code{"data.frame"} or \code{"data.table"}. If this
option is not defined, \code{"data.frame"} will be assumed.
If \code{FALSE}, a \code{data.table} is always returned.}
}
\value{
a data.table with geneset information.
}
\description{
Fetch the active (or all) gene sets from a GeneSetDb or MultiGSEAResult
}
\examples{
gdb <- exampleGeneSetDb()
(gs <- geneSets(gdb))
}

