% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneSetDb-methods.R
\docType{methods}
\name{geneSets}
\alias{geneSets}
\alias{geneSets,GeneSetDb-method}
\title{Fetch the active (or all) gene sets from a GeneSetDb or MultiGSEAResult}
\usage{
geneSets(x, ...)

\S4method{geneSets}{GeneSetDb}(x, active.only = is.conformed(x), ...,
  as.dt = FALSE)
}
\arguments{
\item{x}{Object to retrieve the gene set from, either a \code{GeneSetDb} or a
\code{MultiGSEAResult}.}

\item{active.only}{only look for gene sets that are "active"? Defaults to
\code{TRUE} if \code{x} is conformed to a target expression object, else \code{FALSE}.
\code{\link[=conform]{conform()}} for further details.}

\item{as.dt}{If \code{FALSE} (default), the data.frame like thing that
this funciton returns will be set to a data.frame. Set this to \code{TRUE}
to keep this object as a \code{data.table}}
}
\value{
a data.table with geneset information.
}
\description{
Fetch the active (or all) gene sets from a GeneSetDb or MultiGSEAResult
}
\examples{
gdb <- exampleGeneSetDb()
gs <- geneSets(gdb)
}
