% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-geneSetSelect.R
\name{geneSetSelectUI}
\alias{geneSetSelect}
\alias{geneSetSelectUI}
\alias{gs.render.select.ui}
\alias{gs.select.choices}
\alias{updateGeneSetSelect}
\title{A module that creates a dynamic selectizeInput for a MultiGSEAResult object}
\usage{
geneSetSelectUI(id, label = "Select Gene Set", mg = NULL, server = TRUE)

geneSetSelect(input, output, session, mgc, server = TRUE, maxOptions = Inf,
  sep = "_::_")

updateGeneSetSelect(session, id, label = NULL, choices = NULL,
  selected = NULL, options = list(), server = FALSE)

gs.render.select.ui(ns, choices, server = TRUE, maxOptions = 1000,
  sep = "_::_")

gs.select.choices(mg, sep = "_::_")
}
\arguments{
\item{mg}{\code{MultiGSEAResult} to build options for}

\item{server}{\code{logical} to indicate whether options should be loaded
on the server side (default: \code{TRUE})}

\item{maxOptions}{The maximum number of options to load into the dropdown}

\item{choices}{the output of gs.select.choices(MultiGSEAResult)}

\item{ns}{the namespace function for this module}
}
\value{
a properly wired \code{\link[shiny]{selectizeInput}}

\code{data.frame} to populate \code{choices} of
  \code{selectizeInput}
}
\description{
This code was inspired from the
\href{https://gist.github.com/MarkEdmondson1234/7e56ee7ac5caa74224327489b0849e61}{dynamicSelectShinyModule.R}
gist.

Builds a selectizeInput widget that is specific to a MultiGSEAResult

Note that when returning a data.frame for the choices of a selectizeInput,
we need a column called "value" and a column called "label".
\itemize{
  \item{value}{
    the value that is sent back when an item is selected;
  }
  \item{label}{
    the text that appears in the selection after its triggered
  }
}
}
\section{Module Return}{

Returns information about the geneSetSelect object
}

