% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geneSetSummaryByGenes.R
\docType{methods}
\name{geneSetSummaryByGenes}
\alias{geneSetSummaryByGenes}
\alias{geneSetSummaryByGenes,GeneSetDb-method}
\alias{geneSetSummaryByGenes,MultiGSEAResult-method}
\title{Summarize geneset:feature relationships for specified set of features}
\usage{
geneSetSummaryByGenes(x, features, with.features = TRUE,
  feature.rename = NULL, ...)

\S4method{geneSetSummaryByGenes}{GeneSetDb}(x, features,
  with.features = TRUE, feature.rename = NULL, ..., as.dt = FALSE)

\S4method{geneSetSummaryByGenes}{MultiGSEAResult}(x, features,
  with.features = TRUE, feature.rename = NULL, method = NULL,
  max.p = 0.3, p.col = c("padj", "padj.by.collection", "pval"), ...,
  as.dt = FALSE)
}
\arguments{
\item{x}{\code{GeneSetDb} or \code{MultiGSEAResult}}

\item{features}{a character vector of featureIds}

\item{with.features}{Include columns for \code{features}? If \code{x} is
is a \code{GeneSetDb}, these columns are \code{TRUE}/\code{FALSE}. If
\code{x} is a \code{MultiGSEAResult} object, the values are the logFC of
the feature if present in the gene set, otherwise its \code{NA}.}

\item{feature.rename}{if \code{NULL}, the feature columns are prefixed with
\code{featureId_}, if \code{FALSE}, no renaming is done. If \code{x} is
a \code{MultiGSEAResult}, then this can be the column name found in
\code{logFC(x)}, in which case the value for the feature from the given
column name would be used (setting this to \code{"symbol"}) would be a
common thing to do, for instance.}
}
\value{
a data.frame of geneset <-> feature incidence/feature matrix.
}
\description{
This function creates a geneset by feature table with geneset membership
information for a specified feature set. Only the gene sets that have
any of the \code{features} are included.
}
\examples{
vm <- exampleExpressionSet(do.voom=TRUE)
gdb <- conform(exampleGeneSetDb(), vm)
mg <- multiGSEA(gdb, vm, vm$design, 'tumor', methods=NULL)
features <- c("55839", "8522", "29087")
(gsm.hit <- geneSetSummaryByGenes(gdb, features))
(gsm.fid <- geneSetSummaryByGenes(mg, features, feature.rename=NULL))
(gsm.sym <- geneSetSummaryByGenes(mg, features, feature.rename='symbol'))##'
}
