% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{geneSetURL}
\alias{geneSetURL}
\title{Provides an URL that provides a detailed description for a geneset}
\usage{
geneSetURL(x, i, j, ...)
}
\description{
This function requires that the \code{GeneSetDb@collectionMetadata} slot has
a valid \code{url.fn} column that provides a function to parse the
collection,name information for its genesets into a valid URL. The functions
in \code{url.fn} should take two arguments which correspond to the geneset
\code{collection} and \code{name}, respectively.
}

