% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeggGeneSetDb.R
\name{getKeggGeneSetDb}
\alias{getKeggGeneSetDb}
\title{Retrieves a GeneSetDb from KEGG's REST API}
\usage{
getKeggGeneSetDb(
  species = "human",
  id.type = "entrez",
  database = "pathway",
  ...
)
}
\arguments{
\item{species}{\code{"human"}, \code{"mouse"} or any of the bioconductor or kegg-style
abbreviations.}

\item{id.type}{entrez, enembl?}
}
\value{
A GeneSetDb of the kegg stuffs
}
\description{
Uses \code{\link[limma:getGeneKEGGLinks]{limma::getGeneKEGGLinks()}} and \code{\link[limma:getKEGGPathwayNames]{limma::getKEGGPathwayNames()}}
internally.
}
\details{
Currently we just support the pathway database, and only entrez ids.

Note that \strong{it is your responsibility} to ensure that you can use the KEGG
database according to their licensing requirements.
}
