% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMSigDBset}
\alias{getMSigDBset}
\title{Fetches the requested MSigDB gene sets.}
\usage{
getMSigDBset(id, species = "human", version = .wehi.msigdb.current)
}
\arguments{
\item{id}{A character vector (or integer vector) indicating the desired
genesets}

\item{species}{Character vector indicating which species desired,
\code{"human"} or \code{"mouse"}}

\item{version}{The version of the MSigDB desired. Defaults to latest.}
}
\value{
A list of genesets, named by \code{id}.
}
\description{
Fetches the requested MSigDB gene sets.
}

