% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msigdb-broad.R
\name{getMSigDBset}
\alias{getMSigDBset}
\title{Fetches an MSigDB GeneSetDb}
\usage{
getMSigDBset(collection, species = "human",
  version = .msigdb.version.current)
}
\arguments{
\item{collection}{character vector specifying the collections you want
(c1, c2, ..., c7, h)}

\item{species}{human or mouse?}

\item{version}{the version of the MSigDB database to use.}
}
\value{
a \code{GeneSetDb} object
}
\description{
The gene sets for version 5.0 were parsed straight from MSigDB's XML dumps.
Version 4.0 was based off of the data provided by WEHI:
  http://bioinf.wehi.edu.au/software/MSigDB/
}

