% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{getPantherGeneSetDb}
\alias{getPantherGeneSetDb}
\title{Get pathways/GOslim information from PANTHER.db Biocondcutor package.}
\usage{
getPantherGeneSetDb(type = c("pathway", "goslim"), species = c("human",
  "mouse"))
}
\arguments{
\item{type}{"pathway" or, "goslim"}

\item{species}{"human" or "mouse"}
}
\value{
A wired up GeneSetDb
}
\description{
Get pathways/GOslim information from PANTHER.db Biocondcutor package.
}

