% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-reactome.R
\name{getReactomeGeneSetDb}
\alias{getReactomeGeneSetDb}
\title{Retrieve annotations from reactome.db}
\usage{
getReactomeGeneSetDb(species = "human", rm.species.prefix = TRUE,
  id.type = c("entrez", "ensembl"))
}
\arguments{
\item{species}{the species to get pathay information for}

\item{rm.species.prefix}{pathways are provided with species prefixes from
\code{reactome.db}, when \code{TRUE} (default), these are stripped from the gene set
names.}

\item{id.type}{\code{"entrez"} or \code{"ensembl"}}
}
\value{
a reactome GeneSetDb object
}
\description{
Retrieve annotations from reactome.db
}
