% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.hyperGeometricTest.R
\name{hyperGeometricTest}
\alias{hyperGeometricTest}
\title{Perform Hypergeometric Enrichment tests across a GeneSetDb.}
\usage{
hyperGeometricTest(gsd, selected, universe, direction = c("over", "under"),
  do.conform = TRUE, .external = FALSE)
}
\arguments{
\item{gsd}{The \code{GeneSetDb} object to run tests against}

\item{selected}{The ids of the selected features}

\item{universe}{The ids of the universe}

\item{direction}{Same as direction in \code{GOstats}}

\item{do.conform}{By default \code{TRUE}: does some gymnastics to conform
the \code{gsd} to the \code{universe} vector. This should neber be set
to \code{FALSE}, but this parameter is here so that when this function
is called from the \code{\link{multiGSEA}} codepath, we do not have to
reconform the \code{GeneSetDb} object, because it has already been done.}

\item{.external}{If \code{TRUE} (default), honors the user's setting to
return either a \code{data.frame} or \code{data.table} object. This
setting is defined in the \code{"multiGSEA.df.return"} option
(\code{getOption("multiGSEA.df.return"))}). The value for this option
can be either of \code{"data.frame"} or \code{"data.table"}. If this
option is not defined, \code{"data.frame"} will be assumed.
If \code{FALSE}, a \code{data.table} is always returned.}
}
\value{
A \code{data.table} of results
}
\description{
Perform Hypergeometric Enrichment tests across a GeneSetDb.
}
