% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.hyperGeometricTest.R
\name{hyperGeometricTest}
\alias{hyperGeometricTest}
\title{Perform Hypergeometric Enrichment tests across a GeneSetDb.}
\usage{
hyperGeometricTest(gsd, selected, universe, direction = c("over", "under"),
  do.conform = TRUE, as.dt = FALSE)
}
\arguments{
\item{gsd}{The \code{GeneSetDb} object to run tests against}

\item{selected}{The ids of the selected features}

\item{universe}{The ids of the universe}

\item{direction}{Same as direction in \code{GOstats}}

\item{do.conform}{By default \code{TRUE}: does some gymnastics to conform
the \code{gsd} to the \code{universe} vector. This should neber be set
to \code{FALSE}, but this parameter is here so that when this function
is called from the \code{\link{multiGSEA}} codepath, we do not have to
reconform the \code{GeneSetDb} object, because it has already been done.}

\item{as.dt}{If \code{FALSE} (default), the data.frame like thing that
this funciton returns will be set to a data.frame. Set this to \code{TRUE}
to keep this object as a \code{data.table}}
}
\value{
A \code{data.table} of results
}
\description{
Perform Hypergeometric Enrichment tests across a GeneSetDb.
}
