% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{invalidMethods}
\alias{invalidMethods}
\title{Returns method names that were not run on a MultiGSEAResult}
\usage{
invalidMethods(x, names, as.error = FALSE)
}
\description{
Returns method names that were not run on a MultiGSEAResult
}

