% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-interactive.R
\name{iplot}
\alias{iplot}
\title{Create interactive GSEA plots}
\usage{
iplot(x, y, j, value = c("logFC", "t"), type = c("density", "boxplot",
  "volcano"), tools = c("box_zoom", "box_select", "reset", "save"),
  main = NULL, with.legend = TRUE, with.data = FALSE, ...)
}
\description{
Create interactive GSEA plots
}

