% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{is.active}
\alias{is.active}
\title{Interrogate "active" status of a given geneset.}
\usage{
is.active(x, i, j)
}
\arguments{
\item{x}{\code{GeneSetDb}}

\item{i}{collection of geneset}

\item{j}{name of geneset}
}
\value{
logical indicating if geneset is active. throws an error if geneset
  does not exist in geneSets(x)
}
\description{
This method only works on one geneset at a time (ie. not vectorized)
}

