% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_identifiers.R
\name{load_id_xref}
\alias{load_id_xref}
\title{Load the mouse or human ensembl <-> entrez id maps.}
\usage{
load_id_xref(species = c("human", "mouse"), ..., as.dt = FALSE)
}
\arguments{
\item{species}{do you want the \code{"human"} or \code{"mouse"} xref table?}
}
\value{
a table of ensembl and entrez ids for each species
}
\description{
For convenience, entrez <-> ensembl maps come bundled with the is package.
While there are other sactioned ways to achieve this within the bioconductor
ecosystem, we bundle it here to make our lives easier.
}
\details{
These tables are gnerated by the \code{inst/scripts/genereate-id-maps.R}
}
