% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{logFC}
\alias{logFC}
\title{Extract the individual fold changes statistics for elements in the
expression object.}
\usage{
logFC(x, .external = TRUE)
}
\arguments{
\item{x}{A \code{MultiGSEAResult}}

\item{.external}{If \code{TRUE} (default), honors the user's setting to
return either a \code{data.frame} or \code{data.table} object. This
setting is defined in the \code{"multiGSEA.df.return"} option
(\code{getOption("multiGSEA.df.return"))}). The value for this option
can be either of \code{"data.frame"} or \code{"data.table"}. If this
option is not defined, \code{"data.frame"} will be assumed.
If \code{FALSE}, a \code{data.table} is always returned.}
}
\value{
The log fold change \code{data.table}
}
\description{
Extract the individual fold changes statistics for elements in the
expression object.
}
\examples{
vm <- exampleExpressionSet(do.voom=TRUE)
gdb <- exampleGeneSetDb()
mg <- multiGSEA(gdb, vm, vm$design, 'tumor', methods=NULL)
lfc <- logFC(mg)
}

