% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{melt.gs.scores}
\alias{melt.gs.scores}
\title{Melts the geneset matrix scores from the do.scoreSingleSamples.* methods}
\usage{
\method{melt}{gs.scores}(gdb, scores)
}
\arguments{
\item{gdb}{\code{GeneSetDb} used for scoring}

\item{scores}{The \code{matrix} of geneset scores returned from the various
\code{do.scoreSingleSamples.*} methods.}

\item{a}{melted \code{data.table} of scores}
}
\description{
Melts the geneset matrix scores from the do.scoreSingleSamples.* methods
}
