% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-mgheatmap.R
\name{mgheatmap}
\alias{mgheatmap}
\title{Create a "geneset smart" heatmap.}
\usage{
mgheatmap(gdb, x, col = NULL, aggregate.by = c("none", "ewm", "zscore"),
  split.by = c("none", "geneset", "collection"), rm.dups = FALSE,
  recenter = TRUE, rescale = TRUE, ...)
}
\arguments{
\item{x}{the data matrix}

\item{col}{a colorRamp(2) funciton}

\item{aggregate.by}{the method used to generate single-sample geneset
scores. Default is \code{none} which plots heatmap at the gene level}

\item{split.by}{introduce row-segmentation based on genesets or collections?
Defaults to \code{'none'}}

\item{rm.dups}{if \code{aggregate.by == 'none'}, do we remove genes that
appear in more than one geneset? Defaults to \code{FALSE}}

\item{recenter}{do you want to mean center the heatmap matrix prior to
display?}

\item{rescale}{do you want to rescale the heatmap matrix prior to display?}

\item{...}{parameters to send down to the scoreSingleSample function to
tweak how single-sample geneset scoring is performed.}

\item{the}{\code{GeneSetDb} object that holds the genesets to plot}
}
\description{
Create a "geneset smart" heatmap.
}
