% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-mgheatmap.R
\name{mgheatmap}
\alias{mgheatmap}
\title{Create a "geneset smart" heatmap.}
\usage{
mgheatmap(gdb, x, col = NULL, aggregate.by = c("none", "ewm", "zscore"),
  split = TRUE, scores = NULL, name = NULL, rm.collection.prefix = TRUE,
  rm.dups = FALSE, recenter = TRUE, rescale = TRUE, ...)
}
\arguments{
\item{x}{the data matrix}

\item{col}{a colorRamp(2) funciton}

\item{aggregate.by}{the method used to generate single-sample geneset
scores. Default is \code{none} which plots heatmap at the gene level}

\item{split}{introduce row-segmentation based on genesets or collections?
Defaults is \code{TRUE} which will create split heatmaps based on
collection if \code{aggregate.by= != 'none'}, or based on gene sets
if \code{aggregate.by == "none"}.}

\item{scores}{If \code{aggregate.by != "none"} you can pass in a precomupted
\code{\link{singleSampleGeneSet}} result, otherwise one will be
computed internally. Note that if this is a \code{data.frame} of
pre-computed scores, the \code{gdb} is largely irrelevant (but still
required).}

\item{rm.dups}{if \code{aggregate.by == 'none'}, do we remove genes that
appear in more than one geneset? Defaults to \code{FALSE}}

\item{recenter}{do you want to mean center the rows of the heatmap matrix
prior to calling \code{\link[ComplexHeatmap]{Heatmap}}?}

\item{rescale}{do you want to standardize the row variance to one on the
values of the heatmap matrix prior to calling
\code{\link[ComplexHeatmap]{Heatmap}}?}

\item{...}{parameters to send down to \code{\link{scoreSingleSample}} or
\code{\link[ComplexHeatmap]{Heatmap}}.}

\item{the}{\code{GeneSetDb} object that holds the genesets to plot}
}
\value{
list(heatmap=ComplexHeatmap, matrix=X)
}
\description{
Create a "geneset smart" heatmap.
}
\examples{

vm <- exampleExpressionSet()
gdb <- exampleGeneSetDb()
mgh <- mgheatmap(gdb, vm, aggregate.by='ewm', split=TRUE)
}
