% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-mgTableBrowser.R
\name{mgTableBrowserUI}
\alias{mgTableBrowser}
\alias{mgTableBrowserUI}
\title{A module to explore the GSEA statistics generated from a multiGSEA Run.}
\usage{
mgTableBrowserUI(id)

mgTableBrowser(input, output, session, mgc, method, fdr, server = TRUE)
}
\arguments{
\item{id}{the shiny id of the UI module}

\item{input}{shiny server input object}

\item{output}{shiny server output object}

\item{session}{shiny server session object}

\item{mgc}{the \code{MultiGSEAResultContainer} object}

\item{method}{a reactive that determines which method to explore from this
result}

\item{fdr}{a reactive that gives the fdr threshold to filter results in the
table by.}
}
\description{
The UI is a DT::datatable of GSEA statistics for the selected method. The
module returns a list with the following reactive elements:

\describe{
  \item{$stats}{
    The table of gene sets and their statistics that pass the prescribed
    \code{fdr} thershold
  }
  \item{$selected}{
    The geneset that is currently "active"/selected by the user. This
    is defined as \code{<collection>_::_<name>}
  }
}

You probably want to \code{observeEvent(this$selected)} in your
\code{server.R} (or similar) so you can react to user clicks on different
gene sets
}

