% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb-broad.R
\name{getMSigGeneSetDb}
\alias{getMSigGeneSetDb}
\title{Fetches a \code{GeneSetDb} from geneset collections defined in MSigDB.}
\usage{
getMSigGeneSetDb(collection, species = "human", with.kegg = FALSE,
  version = .msigdb.version.current)
}
\arguments{
\item{collection}{character vector specifying the collections you want
(c1, c2, ..., c7, h)}

\item{species}{human or mouse?}

\item{with.kegg}{The Broad distributes the latest versions of the KEGG
genesets as part of the c2 collection. These genesets come with a
restricted license, so by default we do not return them as part of the
GeneSetDb. To include the KEGG gene sets when asking for the c2
collection, set this flag to \code{TRUE}.}

\item{version}{the version of the MSigDB database to use.}
}
\value{
a \code{GeneSetDb} object
}
\description{
This provides versioned genesets from gene set collections defined in
\href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB}. We strive
to keep the latest version of MSigDB included in this package. The current
version provided is v5.2.

Although the primariy identifiers used in MSigDB are human entrez IDs, we
also include gene symbols, and mouse versions of each gene set (except c1)
that were created from simple orthology mapping via biomaRt.

We also include metadata at the geneset level by adding extra columns
to the \code{geneSets(gdb)} data.frame. "subcategory" may be most useful
for GO gene sets, eg. CC vs BP vs MF, and "organism" defines the organism in
which the geneset was derived from ("Homo sapiens" or "Mus musculus").

Note that this function by default will exclude the KEGG genesets from the
MSigDB c2 collection. Set \code{with.kegg=TRUE} to include them. It is the
end user's responsibility to ensure that they are appropriately adhering
to the licensing restricions of the genesets provided herein.
}
