% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msigdb-broad.R, R/msigdb-wehi.R
\docType{data}
\name{getMSigGeneSetDb}
\alias{MSigDB}
\alias{getMSigDBset}
\alias{getMSigGeneSetDb}
\title{Fetches an MSigDB GeneSetDb}
\usage{
getMSigGeneSetDb(collection, species = "human",
  version = .msigdb.version.current)

getMSigDBset(collection, species = "human",
  version = .msigdb.version.current)
}
\arguments{
\item{collection}{character vector specifying the collections you want
(c1, c2, ..., c7, h)}

\item{species}{human or mouse?}

\item{version}{the version of the MSigDB database to use.}
}
\value{
a \code{GeneSetDb} object
}
\description{
The gene sets for version 5.0 were parsed straight from MSigDB's XML dumps.
Version 4.0 was based off of the data provided by WEHI:
  http://bioinf.wehi.edu.au/software/MSigDB/

This data is downloaded from the
\href{http://bioinf.wehi.edu.au/software/MSigDB/}{WEHI MSigDB for R page},
Which includes \code{*.RData} objects for human and mouse of the
\href{http://www.broadinstitute.org/gsea/msigdb/}{MSigDB database}.
}
\details{
These datasets are versioned, and the last two datasets are kept available
inside this package (starting from v4.0, which was last modified on
21 October 2013). For all analyses using these pre-generated genesets,
the default option is to use the latest version of this data.

The \code{*.RData} objects are simply named lists of Entrez IDs. The names
of the elements of the list correspond to their gene set ID, ie. "chr5q23"
or "KEGG_GLYCOLYSIS_GLUCONEOGENESIS". Each data object holds only one
such list.
}

