% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb-broad.R
\name{getMSigGeneSetDb}
\alias{getMSigGeneSetDb}
\title{Fetches a \code{GeneSetDb} from geneset collections defined in MSigDB.}
\usage{
getMSigGeneSetDb(collection, species = "human",
  version = .msigdb.version.current)
}
\arguments{
\item{collection}{character vector specifying the collections you want
(c1, c2, ..., c7, h)}

\item{species}{human or mouse?}

\item{version}{the version of the MSigDB database to use.}
}
\value{
a \code{GeneSetDb} object
}
\description{
This provides versioned genesets from gene set collections defined in
\href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB}. The following
versions are included in this package:

\itemize{
  \item v5.1
  \item v5.2
}

Starting from version 5.1, the \code{GeneSetDb} includes the symbols for
the features in the genesets (in \code{gdb@db}) as well as additional meta
information about the genesets (in \code{geneSets(gdb)}) for "subcategory"
(which may be most useful for GO gene sets, eg. CC vs BP vs MF) and
"organism" which defines the organism in which the geneset was derived
from ("Homo sapiens" or "Mus musculus").
}
