% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{multiGSEA}
\alias{multiGSEA}
\title{Performs several GSEA analysis over an experiment}
\usage{
multiGSEA(gsd, x, design = NULL, contrast = NULL, methods = NULL,
  feature.min.logFC = 1, feature.max.padj = 0.1, trim = 0.1, ...)
}
\arguments{
\item{gsd}{The \code{\link{GeneSetDb}} to run enrichment over.}

\item{x}{An ExpressoinSet-like object}

\item{design}{A design matrix for the study}

\item{contrast}{The contrast of interest to analyze}

\item{methods}{A character vector indicating the GSEA methods you want to
run. This can be any combination of:
\enumerate{
  \item{camera} {A \code{\link[limma]{camera}} analysis}
  \item{roast} {A \code{\link[limma]{roast}} analysis}
  \item{gsd} {A \code{\link[limma]{geneSetTest} analysis}}
  \item{hyperGeometricTest}{
    Tests for enrichment of differentially expressed features in each
    geneset. Thresholds of differential expression for the individual
    features of a geneset are defind by the \code{feature.min.logFC} and
    \code{feature.max.padj} parameters.
  }
}
If no methods are specified, then only geneset level statistics, such as
the JG score and the mean logFCs t-statistics, of the features in each
geneset are calculated.}

\item{feature.min.logFC}{The minimum logFC required for an individual
feature (not geneset) to be considered differentialy expressed. Used in
conjunction with \code{feature.max.padj} primarily for summarization
of genesets (by \code{\link{geneSetsStats}}), but can also be
used by GSEA methods that require differential expression calls at the
individual feature level, like \code{hyperGeometricTest}.}

\item{feature.max.padj}{The maximum adjusted pvalue used to consider an
individual feature (not geneset) to be differentially expressed. Used in
conjunction with \code{feature.min.logFC}.}

\item{trim}{The amount to trim when calculated trimmed \code{t} and
\code{logFC} statistics for each geneset. This is passed down to the
\code{\link{geneSetsStats}} function.}

\item{...}{The arguments are passed down into the various geneset analysis
  functions.}
}
\value{
A \code{MultiGSEAResult}
}
\description{
multiGSEA is wrapper function which delegates GSEA analyses to different
"workers" that implement each of the GSEA methods specified in the
\code{methods} argument. By default, this only gathers geneset level
statistics without running any of the GSEA methods in particular.
}
\details{
Currently this primarily wraps the GSEA functions from limma, but others
will be added over time. Refer to the \code{method} documentation to see
which GSEA methods are supported.

logFC statistics are always calcualted via a limma pipeline, so if \code{x}
is a \code{DGEList}, it is first "voom"ed, then processed.

This function will write several results to an output directory if
  a valid value for \code{outdir} is provided. Furthermore, the results of
  these analysis will be cached so that time is saved on subsequent calls.
}
\seealso{
\code{\link[limma]{camera}},
  \code{\link[limma]{roast}},
  \code{\link[limma]{geneSetTest}}
}

