% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiGSEA}
\alias{multiGSEA}
\title{Runs several GSEA methods over a dataset for a single contrast.}
\usage{
multiGSEA(x, gene.sets, design = NULL, contrast = NULL,
  methods = c("camera"), outdir = NULL, plots.generate = FALSE,
  plots.padj.threshold = 1, species = NULL, mc.cores = 1L,
  use.cache = FALSE, force.reeval = FALSE, keep.outdir.onerror = TRUE,
  ...)
}
\description{
Currenty supports the following methods:
}
\details{
\enumerate{
  \item \code{limma::camera}
  \item \code{limma::mroast}
  \item \code{limma::geneSetTest}
}

This function will write several results to an output directory
}

