% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MultiGSEAResult-class}
\alias{.MultiGSEAResult}
\alias{MultiGSEAResult-class}
\title{An object to hold the results from a multiGSEA call}
\description{
This is a WIP and we aren't using this yet, but will soon.
}
\section{Slots}{

\describe{
\item{\code{gsd}}{The \linkS4class{GeneSetDb} this analysis was run over}

\item{\code{results}}{The list of individual results generated by each of the
GSEA methods that were run.}

\item{\code{logFC}}{The differential expression statistics for each individual
feature measured in the experiment.}

\item{\code{outdir}}{The directory the images/cache/etc. are saved in.}
}}

