% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{p.matrix}
\alias{p.matrix}
\title{Assembles a matrix of nominal or adjusted pvalues from a multiGSEA result}
\usage{
p.matrix(x, names = resultNames(x), pcol = c("padj", "padj.by.collection",
  "pval"))
}
\arguments{
\item{x}{The output from multiGSEA}

\item{pval}{Are we testing pvalues or adjusted pvalues?}
}
\value{
A matrix of the desired pvalues for all genesets
}
\description{
Assembles a matrix of nominal or adjusted pvalues from a multiGSEA result
}

