% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_identifiers.R
\name{remap_identifiers}
\alias{remap_identifiers}
\title{Converts the feature identifiers in a GeneSetDb to a set of new ones.}
\usage{
remap_identifiers(
  x,
  xref,
  original_id = colnames(xref)[1L],
  target_id = colnames(xref)[2L],
  ...
)
}
\arguments{
\item{x}{The GeneSetDb with identifiers to convert}

\item{xref}{The cross referencing data.frame}
}
\value{
a remapped GeneSetDb object
}
\description{
The various GeneSetDb data providers limit the identifier types that they
provide. Use this function to map the given identifiers to whichever type
you like.
}
\details{
You need to provie a data.frame that has a column for the current identifiers
and another column for the target identifiers.

When there are multiple target id's for the source id, they will all be
returned. When there is no target id for the source id, the soure feature
will be axed.

By default, the first column of the \code{xref} data.frame is assumed to be
the source identifiers, and the second is the target_id you want to remap
the identifiers to.
}
\examples{
xref.fn <- system.file("extdata", "identifiers",
                       "human-entrez-ensembl.csv.gz",
                       package = "multiGSEA")
xref <- data.table::fread(xref.fn, colClasses = "character")
gdb.entrez <- exampleGeneSetDb()
gdb.ens <- remap_identifiers(gdb.entrez, xref,
                             original_id = "entrezgene_id",
                             target_id = "ensembl_gene_id")
}
