% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{renderGeneSetStatsDataTable}
\alias{renderGeneSetStatsDataTable}
\title{Creates a datatable to display gene level statistics for a geneset.}
\usage{
renderGeneSetStatsDataTable(gstats, name, digits = 3,
  feature.link.fn = NULL)
}
\value{
a datatable of gene level statistics
}
\description{
Creates a datatable to display gene level statistics for a geneset.
}
\examples{
vm <- exampleExpressionSet()
gdb <- exampleGeneSetDb()
mg <- multiGSEA(gdb, vm, vm$design)
geneSet(mg, 'c2', 'BURTON_ADIPOGENESIS_PEAK_AT_2HR') \%>\%
  renderGeneSetStatsDataTable('BURTON_ADIPOGENESIS_PEAK_AT_2HR')
}

