% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{renderGseaResultTableDataTable}
\alias{renderGseaResultTableDataTable}
\title{Prepares the datatable of GSEA result statistics for presentation}
\usage{
renderGseaResultTableDataTable(x, method, mg, digits = 3)
}
\arguments{
\item{x}{The set of GSEA statistics generated from from
\code{\link{constructGseaResultTable}}}

\item{mg}{The \code{MultiGSEAResult} object}
}
\value{
a DT::DataTable
}
\description{
Prepares the datatable of GSEA result statistics for presentation
}

