% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{result}
\alias{result}
\title{Extract the result from a given enrichment test from a MultiGSEAResult}
\usage{
result(x, name, stats.only = FALSE, rank.by = c("pval", "t", "logFC"),
  add.suffix = FALSE)
}
\arguments{
\item{x}{MultiGSEAResult}

\item{name}{the names of the results desired}

\item{stats.only}{logical, set to \code{FALSE} if you want to return all
(column-wise) data for each result. By default only the pvalues,
adjusted pvalues, and rank are returned.}

\item{rank.by}{the statistic to use to append a \code{rank} column for the
geneset result. By default we rank by pvalue calculated by the GSEA
method. You can rank the results based on the trimmed mean of the logFC's
calculated for all of the features in the geneset (\code{"logFC"}), or the
trimmed t-statistics of the these features (\code{"t"})}

\item{add.suffix}{If \code{TRUE}, adds \code{.name} as a suffix to the
  columns of the \code{method}-specific statistics returned, ie. the
  \code{pval} column from the \code{camera} result will be turned to
  \code{pval.camera}.}
}
\value{
a data.table with the results from the requested method.
}
\description{
TODO: Enable a way for caller to get a subset of the genesets tested,
perhaps this will happen at the "collectoin" level. In this case, it's
not clear if the pval correction should be redone to only account for
the genesets that are returned from.
}

