% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{resultNames}
\alias{resultNames}
\title{Fetch names of GSEA methods taht were run}
\usage{
resultNames(x)
}
\arguments{
\item{x}{A MultiGSEAResult}
}
\description{
Fetch names of GSEA methods taht were run
}

