% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGSEA.R
\name{results}
\alias{results}
\title{A summary of one/some/all of the results that were run.}
\usage{
results(x, names = resultNames(x), stats.only = TRUE, rank.by = c("pval",
  "logFC", "t"), add.suffix = length(names) > 1L)
}
\arguments{
\item{x}{\code{MultiGSEAResult}}

\item{names}{The results you want to cbind together for the output}

\item{stats.only}{logical, set to \code{FALSE} if you want to return all
(column-wise) data for each result. By default only the pvalues,
adjusted pvalues, and rank are returned.}

\item{add.suffix}{If \code{TRUE}, adds \code{.name} as a suffix to the
columns of the \code{method}-specific statistics returned. This is forced
to \code{TRUE} if \code{length(names) > 1L}}
}
\value{
a data.table with all of the results. The results will optionally
  be suffixed with the method name that generated them when
  \code{length(names) > 1L}
}
\description{
If the result from more than one method is requested, the column names for
these results will be suffixed with the method name.
}

