% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGSEAResult-methods.R
\name{resultNames}
\alias{result}
\alias{resultNames}
\alias{results}
\alias{tabulateResults}
\title{Interrogate the results of a multiGSEA analysis stored in a MultiGSEAResult}
\usage{
resultNames(x)

result(x, name, stats.only = FALSE, rank.by = c("pval", "t", "logFC"),
  add.suffix = FALSE, .external = TRUE)

results(x, names = resultNames(x), stats.only = TRUE, rank.by = c("pval",
  "logFC", "t"), add.suffix = length(names) > 1L, .external = TRUE)

tabulateResults(x, names = resultNames(x), max.p = 0.3, p.col = c("padj",
  "padj.by.collection", "pval"))
}
\arguments{
\item{x}{A \code{MultiGSEAResult} object.}

\item{name}{the names of the results desired}

\item{stats.only}{logical, set to \code{FALSE} if you want to return all
(column-wise) data for each result. By default only the pvalues,
adjusted pvalues, and rank are returned.}

\item{rank.by}{the statistic to use to append a \code{rank} column for the
geneset result. By default we rank by pvalue calculated by the GSEA
method. You can rank the results based on the trimmed mean of the logFC's
calculated for all of the features in the geneset (\code{"logFC"}), or the
trimmed t-statistics of the these features (\code{"t"}).}

\item{add.suffix}{If \code{TRUE}, adds \code{.name} as a suffix to the
columns of the \code{method}-specific statistics returned, ie. the
\code{pval} column from the \code{camera} result will be turned to
\code{pval.camera}.}

\item{names}{The results you want to cbind together for the output, will
default to all of them.}

\item{max.p}{The maximum padj value to consider a result significant}

\item{p.col}{use padj or padj.by.collection?}

\item{x}{MultiGSEAResult}
}
\value{
a data.table with the results from the requested method.

a data.table that summarizes the significant results per method
  per collection for the GSEA that was run
}
\description{
The \code{resultNames}, \code{result}, and \code{results} functions enable
you to explore the results of the analysis run with \code{\link{multiGSEA}}.

The results that are stored within a \code{MultiGSEAResult} object have a
more or less 1:1 mapping with the values passed as \code{methods}, parameter
of the \code{\link{multiGSEA}} call.

The product of an indivdual GSEA is consumed by the corresponding
\code{do.<METHOD>} function and converted into a data.table of results that
is internally stored.

Use the \code{resultNames} function to identify which results are available
for interrogation. The \code{result} function returns the statistics of
one individual result, and the \code{results} function combines the results
from the specified methods into an arbitrarily wide data.table with
\code{method}-suffixed column names.

Use the \code{tabulateResults} function to create a summary table that
tallies the number of significant genesets per collection, per method at
the specified FDR thresholds.
}
\examples{
## Refer to the examples in ?multiGSEA
}

