% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{ret.df}
\alias{ret.df}
\title{Returns data.table objects in formats people like.}
\usage{
ret.df(x, .external = TRUE)
}
\description{
Apparently some people don't like getting data.table(s). Use
\code{options(multiGSEA.df.return='data.frame')} if you want to get
a \code{data.frame}
}

