% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{roundDT}
\alias{roundDT}
\title{Round the numeric columns of a DT}
\usage{
roundDT(x, digits = 3)
}
\arguments{
\item{digits}{the number of digits to round. If \code{NA}, then no rounding
is performed}

\item{a}{DT::datatable}
}
\value{
a rounded DT::datatable
}
\description{
Round the numeric columns of a DT
}

