% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{scoreSingleSamples}
\alias{scoreSingleSamples}
\title{Score each genes over the columns of an expression matrix.}
\usage{
scoreSingleSamples(gdb, y, methods = "ssgsea", melted = FALSE,
  drop.sd = 1e-04, ...)
}
\arguments{
\item{gdb}{A GeneSetDb}

\item{y}{An expression matrix to score genesets against}

\item{methods}{A character vector of methods to score samples by}

\item{melted}{Return a melted data.frame of the results?}

\item{drop.sd}{Genes with a standard deviation across columns in \code{y}
with a standard deviation less than this value will be dropped.}
}
\value{
\code{matrix} with as many rows as \code{geneSets(gdb)} and
  as many columns as \code{ncol(x)}
}
\description{
These functions provide gene set scores at the sample level. No need for
designs and contrasts.
}
\details{
Current scoring methods include:

\enumerate{
  \item zscore: features in geneset rowwise z transformed, then normalized
        by size of geneset (or sqrt(size))
  \item gsva: GSVA package
  \item plage: from GSVA package
  \item ssgsea: from GSVA package
  \item gsdecon: Jason Hackeny's method (kind of like plage), but not
        really.
  \item svd: Simple score that (should) mimic Jason's SVD/eigengene score.
        This is included here mainly to facilitate use of this scoring
        method for people who would have a hard time installing GSDecon2
}
}

