% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{scoreSingleSamples}
\alias{scoreSingleSamples}
\title{Generates single sample gene set scores across a datasets by many methods}
\usage{
scoreSingleSamples(gdb, y, methods = "ewm", as.matrix = FALSE,
  drop.sd = 1e-04, verbose = FALSE, recenter = FALSE,
  rescale = FALSE, ..., as.dt = FALSE)
}
\arguments{
\item{gdb}{A GeneSetDb}

\item{y}{An expression matrix to score genesets against}

\item{methods}{A character vector that enumerates the scoring methods you
want to run over the samples. Please reference the "Single Sample Scoring
Methods" section for more information.}

\item{as.matrix}{Return results as a list of matrices instead of a melted
data.frame? Defaults to \code{FALSE}.}

\item{drop.sd}{Genes with a standard deviation across columns in \code{y}
that is less than this value will be dropped.}

\item{verbose}{make some noise? Defaults to \code{FALSE}.}

\item{recenter, rescale}{If \code{TRUE``, the scores computed by each method are centered and scaled using the }scale\code{function. These variables correspond to the}center\code{and}scale\code{parameters in the}scale\code{function. Defaults to}FALSE`.}

\item{...}{these parameters are passed down into the the individual single
sample scoring funcitons to customize them further.}

\item{as.dt}{If \code{FALSE} (default), the data.frame like thing that
this funciton returns will be set to a data.frame. Set this to \code{TRUE}
to keep this object as a \code{data.table}}
}
\value{
A long data.frame with sample,method,score values per row. If
\code{as.matrix=TRUE}, a matrix with as many rows as \code{geneSets(gdb)}
and as many columns as \code{ncol(x)}
}
\description{
It is common to assess the activity of a gene set in a given sample. There
are many ways to do that, and this method is analogous to the
\code{\link[=multiGSEA]{multiGSEA()}} function in that it enables the user to run a multitude of
single-sample-gene-set-scoring algorithms over a target expression matrix
using a \code{\link[=GeneSetDb]{GeneSetDb()}} object.
}
\details{
Please refer to the "Generating Single Sample Gene Set Scores" of the
multiGSEA vignette for further exposition.
}
\examples{
library(reshape2)
gdb <- exampleGeneSetDb()
vm <- exampleExpressionSet()
scores <- scoreSingleSamples(gdb, vm, methods=c('ewm', 'ssgsea', 'zscore'),
                             uncenter=FALSE, unscale=FALSE,
                             ssgsea.norm=TRUE)
sw <- dcast(scores, name + sample ~ method, value.var='score')
corplot(sw[, c("ewm", "ssgsea", "zscore")],
        title='Single Sample Score Comparison')
}
