% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{scoreSingleSamples}
\alias{scoreSingleSamples}
\title{Score each genes over the columns of an expression matrix.}
\usage{
scoreSingleSamples(gdb, y, methods = "ssgsea", melted = FALSE, ...)
}
\arguments{
\item{gdb}{A GeneSetDb}

\item{y}{An expression matrix to score genesets against}

\item{score.fn}{The function used to summarize the genes in a geneset.
If left \code{NULL}, defaults to \code{mean}}

\item{trim}{The amount to trim in the trimmed mean of the individual
geneset genes/features.}
}
\value{
\code{matrix} with as many rows as \code{geneSets(gdb)} and
  as many columns as \code{ncol(x)}
}
\description{
These functions provide gene set scores at the sample level. No need for
designs and contrasts.
}
\details{
Current scoring methods include:

\enumerate{
  \item zscore: features in geneset rowwise z transformed, then normalized
        by size of geneset (or sqrt(size))
  \item gsva: GSVA package
  \item plage: from GSVA package
  \item ssgsea: from GSVA package
  \item gsdecon: Jason Hackeny's method (kind of like plage), but not
        really.
  \item svd: Simple score that (should) mimic Jason's SVD/eigengene score.
        This is included here mainly to facilitate use of this scoring
        method for people who would have a hard time installing GSDecon2
}
}

