% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{scoreSingleSamples}
\alias{scoreSingleSamples}
\title{Score each genes over the columns of an expression matrix.}
\usage{
scoreSingleSamples(gdb, y, methods = "plage", melted = FALSE, ...)
}
\arguments{
\item{gdb}{A GeneSetDb}

\item{y}{An expression matrix to score genesets against}

\item{score.fn}{The function used to summarize the genes in a geneset.
If left \code{NULL}, defaults to \code{mean}}

\item{trim}{The amount to trim in the trimmed mean of the individual
geneset genes/features.}
}
\value{
\code{matrix} with as many rows as \code{geneSets(gdb)} and
  as many columns as \code{ncol(x)}
}
\description{
Score each genes over the columns of an expression matrix.
}

