% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSingleSamples.R
\name{ssGSEA.normalize}
\alias{ssGSEA.normalize}
\title{Normalize a vector of ssGSEA scores in the ssGSEA way.}
\usage{
ssGSEA.normalize(x)
}
\arguments{
\item{x}{a \code{numeric} vector of ssGSEA scores for a single signature}
}
\value{
normalized \code{numeric} vector of \code{x}
}
\description{
Normalize a vector of ssGSEA scores in the ssGSEA way.
}

