% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GeneSetDb-methods.R
\name{subset.GeneSetDb}
\alias{subset.GeneSetDb}
\title{Subset GeneSetDb to only include specified genesets.}
\usage{
\method{subset}{GeneSetDb}(x, keep)
}
\arguments{
\item{x}{\code{GeneSetDb}}

\item{keep}{logical vector as long as
\code{nrow(geneSets(x, active.only=FALSE))}}
}
\value{
a \code{GeneSetDb} that has only the results for the specified
  genesets.
}
\description{
This isn't exported yet because I don't like its implementation
}

