% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{subsetByFeatures}
\alias{subsetByFeatures}
\title{Subset a GeneSetDb to only include geneSets with specified features.}
\usage{
subsetByFeatures(x, features, value = c("featureId", "x.id", "x.idx"), ...)
}
\arguments{
\item{x}{\code{GeneSetDb}}

\item{featureIds}{Character vector of featureIds}
}
\value{
A subset of \code{x} which contains only the geneSets that contain
features found in \code{featureIds}
}
\description{
Subset a GeneSetDb to only include geneSets with specified features.
}
\examples{
gdb <- exampleGeneSetDb()
features <- c("55839", "8522", "29087")
(gdb.sub <- subsetByFeatures(gdb, features))
}
