% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarizeResults}
\alias{summarizeResults}
\title{Creates a summary table from all the methods that were run}
\usage{
summarizeResults(x, design, contrast, gst, results, logFC = NULL,
  score.by = c("logFC", "t"))
}
\description{
Creates a summary table from all the methods that were run
}

