% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-sample-scoring-methods.R
\name{svdScore}
\alias{svdScore}
\title{Calculates the (optinally) expression-rescaled eigengene for each sample.}
\usage{
svdScore(x, eigengene = 1L, center = TRUE, scale = FALSE,
  uncenter = FALSE, unscale = FALSE, retx = FALSE)
}
\arguments{
\item{x}{matrix of genes x samples}

\item{eigengene}{the "eigengene" you want to get the score for. only accepts
a single value for now.}

\item{center}{center the data first?}

\item{scale}{scale the data first?}

\item{uncenter}{uncenter the data on the way out?}

\item{unscale}{the data on the way out?}

\item{retx}{Works the same as `retx` from \code{\link[stats]{prcomp}}. If
\code{TRUE}, will return a \code{ret$pca$x} matrix that has the rotated
variables.}
}
\value{
list of useful transformation information
}
\description{
This function returns the innards of a PCA over "sample space"
}

