% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-sample-scoring-methods.R
\name{svdScore}
\alias{svdScore}
\title{Calculates the (optinally) expression-rescaled eigengene for each sample.}
\usage{
svdScore(x, center = TRUE, scale = FALSE, uncenter = center,
  unscale = scale)
}
\arguments{
\item{x}{matrix of genes x samples}
}
\description{
This function returns the innards of a PCA over "sample space"
}

