% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Create an interactive volcano plot}
\usage{
volcano_plot(x, stats = "dge", xaxis = "logFC", yaxis = "pval", idx,
  highlight_genes = NULL, xtfrm = identity, ytfrm = function(vals)
  -log10(vals), horiz_lines = c(padj = 0.1), xgran = NULL, ygran = NULL)
}
\arguments{
\item{highlight_genes}{A vector of featureIds to highlight, or a GeneSetDb
that we can extract the featureIds from for this purpose.}
}
\description{
Create an interactive volcano plot
}

