% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-sample-scoring-methods.R
\name{zScore}
\alias{zScore}
\title{Calculate geneset score by average z-score method}
\usage{
zScore(x, summary = c("mean", "sqrt"), trim = 0)
}
\arguments{
\item{x}{gene x sample matrix}

\item{summary}{sqrt or mean}

\item{trim}{calculate trimmed mean?}
}
\description{
Calculate geneset score by average z-score method
}
